Enable vr mode and persistent vr mode at device boot
1. Enable vr mode and persistent vr mode at device boot for "boot to vr"
devices.

2. When a boot to vr device wakes up from sleep, immediately reenter vr
mode and persistent vr mode.

One unexpected change I had to make was to move when we're allowed to
enter vr mode from the "third party apps can start" boot phase to after
the device is unlocked. Without that change, when we try to connect to
the vr service on boot to vr devices the service manager code was
rejecting our connection request because our vr listener service isn't
direct boot aware (specified with directBootAware="true" in the service
declaration in the manifest). Waiting until the device is unlocked fixes
this issue.

Bug: 35627160

Test: - Marlin: Confirmed device boots with vr mode off.

- Marlin: Confirmed basic app behavior in VrHome and cube sea, in both
  the old and new render paths.

- Marlin: Confirmed we exit vr mode and persistent vr mode when the
  power button is pressed.

- Marlin: Ran the above tests with a lock screen and again without a
  lock screen, and confirmed correct behavior in both cases. This is
  relevant because of the switch to allow vr mode only once the device
  is unlocked.

- Boot to vr device: Confirmed at boot that vr mode and persistent vr
  mode are entered immediately after the device is booted, with no vr
  component. Once VrHome starts the vr component switch is successfully
  communicated to the vr listener service.

- Boot to vr device: Confirmed we exit vr mode and persistent vr mode
  when the power button is pressed, and immediately reenter vr mode and
  persistent vr mode when the power button is pressed again.

Change-Id: I9569f763f03586f81ab531b1460688d6f82f5341