Add android.test.mock to the optional runtime libraries
This is needed in order to make the following manifest entry work
properly.
  <uses-library android:name="android.test.mock"/>

Tested by adding the entry to an APK and installing it which caused an
error about an unknown library. Then updated this file, rebuilt,
reflashed and installed APK again at which point it worked.

Follow up change to cts/tests/signature will add proper test for this
library.

Bug: 30188076
Test: see above
Change-Id: I630b7bc48a50ab1c52bb5feed54c2e4deb876339