Don't reset review state until no permission requires a review.
In a permission review mode we have a per user flag whether a
permission review is required for an app to avoid iteration
over all permissions on an activity start. This state was
incorrectly cleared if a permission review flag is cleared
despite the fact that there may me other permissions requiring
a review.

Test: manual

bug:63720758

Change-Id: I3391c9d7e419e2e22f7b053de8e6ed977b6eba91