Fix system server compilation in ZygoteInit
The code which dexopt SystemServer in ZygoteInit used to pass the
sharedLibrary string in the old format.

This CL updates the classPath string to the new class loader context
encoding.

Bug: 63941300
Test: compile WITH_DEXPREOPT=false and WITH_DEXPREOPT_BOOT_IMG_ONLY=false
and check that when the device boots the correct context is passed to
dex2oat.

(cherry picked from commit b0dede3ec2e57a734fe17bef4920020c2d66488c)

Change-Id: Ie27cd21b0b9c8b7e226bce74dfa297569ac12317