Remove TODO related to WakeLocks
There's no need to keep a WakeLock in the RulesManager - if
the device goes to sleep it should continue the work when
it wakes up again. All I/O is handled via AtomicFile and should
be robust to corruption, etc. and this reduces the likelihood
of issues caused by sudden sleeps.

Bug: 31008728
Test: make
Change-Id: I3d5076ca7d2f39076f76bdb745cedde214eac21b