Correct missed use of work paint for hyphenation
The change in I0b95a10c05711350950d6c8d6501f6af9414d413 copied the
Paint to a work paint to make sure we don't modify it, but missed
actually using that Paint, resulting in hyphenation basically getting
ignored in getLineExtent().

Now we correctly pass the work paint to the TextLine used.

Change-Id: Ida89ba2b11d35d93012abb0adf9bb29d1b09c2d2
Fixes: 64998667
Bug: 33388205
Bug: 64452136
Test: bit FrameworksCoreTests:android.widget.TextViewTest
Test: bit FrameworksCoreTests:android.text.
Test: bit CtsTextTestCases:android.text.cts.