Don't defer client hidden if activity is already paused
We defer client hidden to allow activities that support pip to
enter pip before onPause() returns. However, the condition only
checked for stopping and stopped states. There are situations
where the app is already paused and we are now moving it to the
invisible-stopped state. For does cases we need to make sure not
to defer client hidden, so that the right visibility is reported
to the app and the app can in-turn relayout out its windows to
invisible so that window manager can distroy and surfaces
associated with the windows when the exit animation is done.

Change-Id: Ib1c4cf6a989f597ba700ce460fca4b012bf5146b
Fixes: 63117546
Test: go/wm-smoke