AAPT2: Handle inflating compiled attributes.
* Added the resource ID pulled from the XML parser into the compiled
attribattribue. This allows binary attributes such as the version code
and min SDK values to be available whern flattening the XmlResource back
into an APK file.

* Only add a compiled value if the attribute value was not a string.
This allows string values like the package name and version name to be
flattened back into an APK file.

Before making this change, aapt would segfault when dumping the manifest
and dumping badging would be missing values. After making this change,
the manifest and badging are dumping correctly.

Test: ran unit tests
Test: manually ran optimize against an APK
Change-Id: I885bb06557b032b9f702e8065d637d359c2b519b