Prevent dismiss divider going minimized to recents for other users
The bus event DockedFirstAnimationFrameEvent was not fired to
DividerView from secondary to system user so the last snap target for
DividerView was not recorded instead the bottom of the screen was the
last target. Since the bottom of the screen is last recorded, when
unminimizing, the dividerview goes to the bottom of the screen and
dismisses. This is fixed by piping the event from the secondary to
system user so that DividerView can get the event.

Fixes: 67326180
Test: manual - go to secondary user, drag to dock and minimize, press
recents button
Change-Id: Icd7497c49826d04dc907ef2eff3a5248e78ef67b