Set updated system package's mExtra
While scanning system packages, we abort the process in the
middle of the scan and leave our internal structures in an
ill defined state. This change is not about making the structures
consistent; rather unblocking a build breakage. We'll look
at data consistency during a later refactoring stage when
the process of package scanning is fixed.

A reboot is necessary for the bug to exhibit itself because
the permissions are re-granted [or denied] on every boot.

Change-Id: I96cc2f76623911f2cf93727e9f1787b42210a8d6
Fixes: 68328870
Test: Manual.
Test: Add a new priv-app that take a privieleged permission.
Test: See that the permission is granted.
Test: Update the app and reboot
Test: See that the permission is still granted.
Test: cts-tradefed run commandAndExit cts-dev -m CtsAppSecurityHostTestCases -t android.appsecurity.cts.PermissionsHostTest
Test: cts-tradefed run commandAndExit cts-dev -m CtsPermissionTestCases
Test: cts-tradefed run commandAndExit cts-dev -m CtsPermission2TestCases
Test: bit FrameworksServicesTests:com.android.server.pm.PackageManagerSettingsTests
(cherry picked from commit 815a35926d12b0f7c912e71254cd87f8fc01c45d)