Add ability to customize button corner radius.
Creates a new buttonCornerRadius public attribute to control
the rounded corner radius for buttons at the theme level.

Updates default button backgrounds to use new value and updates
themes to set the corner radius to existing value of 2dp.

Also adds a configuration hook for a device to change the value
for DeviceDefault-based themes.

Bug: 64122537
Bug: 69314526
Bug: 69367197
Test: ag/3238215
Change-Id: I00be0c53b99292e1728ecdf40d4c876e283e65da