Sort "Recent Location Requests" by recency.
The list of apps with recent location requests is currently displayed in
alphabetical order on device (Settings -> Security & Location ->
Location). Sort this list by recency: (1) apps that are currently
requesting location are displayed  on the top. (2) For apps that are not currently
requesting location, the ones with most recent location request finish
time come first.

Bug: 70350519
Test: Robolectric
Test: manual
Change-Id: I5c757defcd7645d254c9c47e3c83f7e323247a71