Keyguard can be on even if not showing
It is possible for keyguard to be in "restricted input" state even if
isKeyguardShowingAndNotOccluded returns false. The previous CL to
deprecate inKeyguardRestrictedInputMode has also changed the behaviour
of isKeyguardLocked. Return to the old behaviour here, but keep the API
deprecated.

Bug: 70411251
Test: 1) add Google account to phone 2) Enable google assistant and set
up active edge to trigger the assistant 3) Ensure phone is in the
always-on-display mode 4) Squeeze the phone to trigger the assistant.
Previously, the assistant would not launch. The expected behaviour is
that assistant gets launched. Without this change, display will turn on
but the assistant will not get launched.

Change-Id: Ic299f3fc223ebdc3b0f6eea97cb4a89a3142e44a
(cherry picked from commit 6e3e49264c1694aaaaf85290ee385ac536c2054d)