Add apis needed for android.test.legacy
The long term goal of removing junit and dependent android.test classes
from the android.jar is the removal of the runtime libraries that
provide those classes. A key part of the strategy for migrating APKs off
the runtime libraries is the android.test.legacy library which APKs can
statically include in place of the runtime dependencies without having
to change their source code.

The longer term goal is for all tests to migrate from these classes
altogther and to use the Android Test Support Library instead but that
will require significant changes to the test source code.

In order for it to be safe to statically include android.test.legacy in
an APK it cannot use any internal APIs. This change adds the additional
classes and methods needed to the android.test.mock API library. The
additions are all marked as deprecated to try and avoid developers
from using them in new tests.

It also marks the AccountManager constructor used by MockAccountManager
with the android.annotation.MockApi so that it is included in the mock
stubs libraries against which the android.test.mock.stubs[-system]
libraries build.

Bug: 30188076
Test: make checkbuild

(cherry picked from commit 89e4cd3ce24bf6388eac7fbb71f66fcde95b46c3)

Merged-In: I85417cc328cab898ab93b9b726648f4232555324
Change-Id: If3b6fc630df93a44e4c07ef769737a2ac3ebf476