Use value from XML for Secure.NOTIFICATION_BADGING.
Config config_notificationBadging was created, and it was used for
showing/hiding the toggle in Settings app. However, since there's no
default value for Settings.Secure.NOTIFICATION_BADGING, it is possible
to be in a state where there's no toggle, and the user is stuck having
it turned on. This change pulls in config_notificationBadging for also
the default value of Secure.NOTIFICATION_BADGING.

Bug: 36021111
Test: Build, flashed and confirmed proper config value. Also don't see
the toggle anymore in Settings app.

Change-Id: Idbf162040d29cf2456356e7c628a29b26fef6363