Refactored out app standby bucket reordering
Added a separate method to reorder alarms when app standby bucket
changes. Using rebatchAllAlarms was not ideal because it removes and
inserts all the alarms again, which may not be needed in most of the
cases relating to app standby.
Excluding allow_while_idle_alarms for now due
to a potential infinite loop when triggerAlarmsLocked and
adjustDeliveryTimes... would keep updating delivery times to values
unacceptable to each other.

Test: atest CtsAlarmManagerTestCases

Bug: 72660630
Bug: 72816079
Change-Id: I5406e14c12e18b50046ef9b39b7710ba9d0383b2