Fixed bug with turnScreenOn window flag.
The test
ActivityManagerActivityVisibilityTests#testTurnScreenOnActivity_withRelayout
was failing due to issues with how the turnScreenOn flag was handled.
This change centralizes the logic about when to turn the screen on. The
check for whether the screen should be turned on will happen on each
relayout but is guarded by a flag set in resume. This way the screen can only
have the chance to turn on if a resume was just called.

Change-Id: I0a6e4a11de92fa41c8a49cbfd1b78bd8b38a1b4a
Fixes: 68091612
Test: WindowStateTests
Test: ActivityManagerActivityVisibilityTests