Fixes home button from moving off its origin position after rotation
It is possible to quick scrub from a portrait app to a forced landscape
app to resultantly move the home button off its original position or
even off the screen. Force animation end to reset the position of the
home button and prevent null exceptions of the home button when running
animation end.

Fixes: 73507167
Bug: 70180755
Test: manual - quick switch/scrub between portrait and landscape forced
apps

Change-Id: Iea247b1f75701a171d1ddfd117e2056c2321c835