Fixed issue with focusing on minimized split-screen stack
If the current split-screen secondary stack is finishing and we are
moving focus to another stack, we should prefer the focus to move to
another split-screen secondary stack or fullscreen stack instead of
the primary split screen stack to avoid:
- Moving the focus to the primary split-screen stack when it can't
be focused because it will be minimized, but AM doesn't know that yet
- primary split-screen stack overlapping with a fullscreen stack when
a fullscreen stack is higher in z than the next split-screen stack.
Assistant stack, I am looking at you...
We only move the focus to the primary-split screen stack if there isn't
a better alternative.

Change-Id: Iffaaa56a88e24bbf0f6303acd924700257baad0b
Fixes: 74404706
Fixes: 70677280
Fixes: 74366688
Test: atest ActivityManagerAssistantStackTests#testAssistantStackLaunchNewTaskWithDockedStack