Update Wcdma and Tdscdma CellInfo for HAL 1.2
In HAL 1.2 WCDMA and TDSCMA were updated to bring
them in line with the other CellInfo types. WCDMA
was missing support for Rscp and Ec/No; the former
being the primary measurement used for WCDMA signal
strength. TDSCDMA was missing support for RSSI, and
the definition for RSCP was incorrect, so it was fixed
to use the PCCPH RSCP.

Furthermore, TDSCDMA support was unavailable except
in the CellIdentity. This patch pipes support from the
HAL to the framework, adding CellInfoTdscdma, which is
primarily for Asia.

-Add RSCP and Ec/No fields to CellSignalStrengthWcdma
-Add CellInfoTdscdma and CellSignalStrengthTdscdma
-Add UARFCN and Alphas support for CellIdentityTdscdma
-Fix a bug in the equals() check for CellIdentity
-Update documentation for CellInfo<RAT> classes and
 the getLevel() methods of CellSignalStrength<RAT>

Bug: 76153768
Test: RILTests added for TDSCDMA
Merged-In: I71e0998f1a9bd7656d495003a8ee5ef616ca4f43
Change-Id: I71e0998f1a9bd7656d495003a8ee5ef616ca4f43
(cherry picked from commit 9c50d93752e8d6e26bc80ab6a743d50686a6729d)