Fixed the touch targets when replying
Because part of the reply field was outside the notification,
clicks on the remote input or the send button could be discarded
and go to the view instead.
We're not manually dispatching those touches to the remoteInput.
Additionally are we now closing the input field first and only
then opening the app.

Change-Id: Iaea3fb78347dfc3a3e22b0d7155e6d2e50c82285
Fixes: 74161213
Fixes: 77482496
Test: open inline reply, click on very bottom below text, observe that app isn't opened