Restore the OEM hook implementation and usage
The current plan is to keep the OEM hook implementation in the 
framework, this is for legacy device and old device upgrading 
(per b/78098059).

For P launching device, the OEM hook service is removed in hal, 
so the OEM hook api in framework should return something dummy, and 
prevent infinite loop to get service, but these parts should be done
in seperate CLs after this restore CL is merged.

This reverts commit fa74edc757dee07ba80f11248102c3d6f495abeb.

Bug: 34344851
Test: pass Treehugger tests; pass cuttlefish test RilE2eTests
Change-Id: I1f2117331a7f846b192c5f956f3841d58cb8154b