WindowStateAnimator: Hide preserved surface when showing new Surface
Our flow of operation when preserving Surfaces has always been:
    1. Open Transaction
    2. Show new surface
    3. Close transaction
    4. Destroy preserved surface outside of transaction

This means there is a small time after 3 where both surfaces are visible.
Normally this isn't a problem because the new surface totally covers up the old
Surface however in cases where we have partial translucency the two surfaces
on top of eachother will not look correct. The easiest fix for now is to hide
the old surface when showing the new one. The better fix for later is to use
destroyInTransaction for cleaning up the preserved surfaces.

Bug: 78639547
Test: Manual
Change-Id: I7dc687cf6131fb3af3881ccd9622e53d72e89e7d