Decrease the inferred auto-brightness adjustment at the extremes.
Now that the brightness slider always represents the brightness of the
panel, rather than the auto-brightness adjustments like it did prior to
P, we need to infer the adjustment by comparing the old and new
brightness values. At the extremes, this inferred brightness increases
dramatically towards (+/-) infinity, so we use a different strategy.
Based on dogfood feedback, the current inference seems to have too
strong of an effect, so we're cutting it in half in the hopes that this
is more reasonable.

Change-Id: If6a6c3355f9dcce8588512f1d969c871149ec299
Fixes: 79153846
Test: manual