WM: Prevent unminimizing docked stack on AOD
When we go to AOD, we hide Home's app token - which causes
the docked stack to unminimize. There was some protection against
that because the same thing happens on the keyguard, but that did
not work because the keyguard state is not aware of AOD. Instead,
we pipe the source of truth from AM down to WM and use this instead.

Change-Id: I68d1079a6bbd9080816ee52c1b5650623a661421
Fixes: 79267409
Test: Enter split screen, go to home, turn of screen, turn on screen. Verify docked stack stays minimized and there is no janky transition