Fix EditText show wrong position when ime show
In landscape and largest density, the area that could show the heads
up notification is smaller than normal density. To handle the heads
up notification should be different with the normal notification in
both of targetScrollForView and getScrollRange because
StackScrollAlgorithm translation heads up notification as
mHeadsUpInset but neither targetScrollForView nor getScrollRange has
have the same consideration.

The solution make the StackScrollAlgorithm to considerate scroll
value to move up the heads up notification.

The getScrollRange need to consider more about the difference
between heads up showing and non-heads up showing. The top padding
and the total content height are the major considerations.

Test: manual test
Test: atest SystemUITests
Change-Id: I46bb960a81f6ab5be4d8e39738e8fb64245dd4eb
Fixes: 110388615