Fix for CTS testAllowWhileIdleJobInTempwhitelist
In this test case when all constraints to run the jobs are satisfied,
then to make job ready to run, intent MSG_CHECK_JOB is send to job handler.
But this intent is not dispatched to job handler.

BUG: 113778907
Test:After these changes this CTS case always passed.

Change-Id: I8c63361dcd7ade19b724c9bc2fdf78d5f10c8d71
Merged-In: I8c63361dcd7ade19b724c9bc2fdf78d5f10c8d71

Feature: CTS(*)