Revert "preloaded-classes: remove core HIDL libs"
This reverts commit d20f0d811f96d3bbc7d79aa73c67e84804de6f09.

Appears to be the cause of a 14ms regression in wifi startup time in system server that could be offset here (and potentially other places) by keeping the preload in zygote. A local trial of five showed this revert to take it 111.2ms -> 108ms. The variance is likely too high for this to be significant, however, logically, in the short term, these libraries are being used, and so preloading makes sense. I'll leave further changes to this list to the owners of this list.

Test: boot walleye 5 times w/ and w/o this test, and find small gain in reverting this.
Reason for revert: b/115771598

Change-Id: Ia53dbe8b70900d6f04011b15cbdddf786378fc3b