Make TIME_TICK an unordered broadcast
TIME_TICK is a foreground broadcast that goes only to registered
receivers, but was being dispatched via the ordered queue as a side-
effect of being sent directly as an alarm operation.  We now use a
direct method-call alarm as the trigger and send it as an unordered
broadcast instead.  This eliminates the potential for observed issues
of slow TIME_TICK handling leading to significant latencies for
subsequent foreground-priority broadcasts such as live phone call
flows.

Bug: 111404343
Bug: 116719392
Test: atest CtsAlarmManagerTestCases
Test: 'dumpsys activity broadcasts' and observe that TIME_TICK
      is no longer being dispatched as "ordered=true"

Change-Id: I9c768668aaf6447221c905969ee6948e81958eb0