Prevent clock becoming visible when icon disabled
When the status bar clock is disabled (via a "clock" entry in icon_blacklist
setting) it could mistakenly reappear due to
CollapsedStatusBarFragment#animateShow (which directly modifies the
visibility of the view via setVisibility). Prevent this by overriding
setVisibility to stop the view becoming visible when it should be
it should be hidden.

Test: am start com.android.systemui/com.android.systemui.DemoMode
      Status bar > Toggle Clock to "don't show"
      (Status bar clock should disappear)
      Pull down shade
      (Status bar clock should remain gone)

Bug: b/113087760 (See 2.)
Change-Id: I8537fec1f29c928cdf21a60c209cfb8ac86cc95a