WM: Track waitForConfig per display
If a virtual display is present while the screen is rotated, reconfiguring the virtual display
will set the global mWaitingForConfig flag and can prevent window relayouts and
surface placements for windows on the primary display. This will cause windows to be
drawn incorrectly in the old rotation and get stuck in this state until there is another rotation.

Bug: b/117770934
Test: go/wm-smoke
Test: Test rotation with huddle presenting
Change-Id: I14034a8e8aac857b46702c0108e07adcb4eb8ac5