Fix wallpaper size changed to -1*-1 on first boot.
In previous version we have ensured the DisplaySize won’t be invalided
when getting it from getDisplayDataOrCreate, The problem occurs when
there is no wallpaper (such as first boot or xml error), it will reset
the size to -1 then tries to restore an image wallpaper. Then it will
go wrong because the bitmap size is set to -1.

Change-Id: I71faa85cbbb5dcbb13414a15b1c9e3ad2a37887b
Fix: 120114212
Test: Manually test back/restore wallpaper.
Test: atest WallpaperManagerTest
Test: atest WmTests