Only log smart suggestions when their notification has been expanded.
Smart suggestions in notifications are only shown when the notification
is expanded. Before this CL we logged the SMART_REPLY_VISIBLE event the
first time the notification is visible - with this CL we will log that
event the first time the notification is visible AND expanded at the
same time (i.e. when smart suggestions are visible).

Bug: 120967525
Test: atest NotificationManagerServiceTest
Test: manual - using logcat, ensure we only log smart suggestions when
they are actually shown (when the notification is expanded and visible,
but not when only expanded or only visible).

Change-Id: Ic435cf3c33db770095d903c72c0016241b1549e1