Revert "[framework] Compile secondary dex files in isolation"
This reverts commit 41df032fdf6b6c976ec80fbe1f99bc3b978e35a7.
ART does not compile secondary dex files in the user process
any more. The original issue of a large overhead of providing
a class loader context is therefore not relevant any more and
we can start passing the context to installd.

Note that this also partially reverts commit
532428872b241f92497ae0b4fec24950fe25de0c which restricted
visibility of methods in PackageDexUsage. The commit also
removed data collection for unsupported class loaders. The
logic in PackageDexOptimizer was adjusted accordingly.

Bug: 64530081
Bug: 111174995
Test: manual
Change-Id: Id78b5a6d8841b199c12a63a8d45d12efbcc32275