Clarify shell/root users also have DUMP permission
This is a follow up CL to my previous CL [1], which aimed to
streamline and centralize when and how caller verifications are
performed inside InputMethodManagerService#onShellCommand().

The goal of this change is just to make it clear that shell/root users
also have DUMP permission.  There should be no behavior change as long
as those users have DUMP permission because IMMS#onShellCommand() is
already locked down to accept IPCs only from these two UIDs [2].

The point is that ShellCommand#handleDefaultCommands() by default
dispatches 'dump' command to Binder#doDump() method.  Thus it would
make sense to clarify that DUMP permission is also granted to
shell/root users for completeness.

 [1]: I6fd47b5cc1e7da7222774df20247a2c69a70f45b
      8f5ea96fe4709a309d904102cc402701a34b12ac
 [2]: If87189563ccaacd4f9c666bab4f9ad08a9343084
      f9ca7cc415ec56a0ec6e70178f5cf5113326f715

Bug: 34886274
Test: 'adb shell cmd input_method dump' still work
Change-Id: If91137e0c144d90ac3046f30607e36bb957a7e0e