Added AccessPoint cache integration for OSU AccessPoints
Cached OSU APs will be used instead of creating a new OSU AP on every
updateAccessPoints(). This will ensure that the OSU AP seen by
WifiTracker is the same as the OSU AP passed to the
AccessPointPreference, allowing the RSSI to be updated while retaining
information about the OSU provisioning status and failure code.

Tracking bug for adding robolectric tests: b/122849296

Bug: 118705403
Test: manual, build and log AP creation and cache access.
Change-Id: I0c3fdbeb2a28b5c4005a11ed1650f0c40a4f1d5c