Revert "Select only preinstalled Spell Checker Services"
This reverts commit 2df11d3c80ab179d7120311f661799e7003d88f2.

Reason for revert: The backport for b/118694079 was applied improperly.

The fix involved filtering the class members list of spell checkers into a new list before searching it. The backport filtered the list but failed to update references to the class member into references to the local variable, creating no change in observable behavior. A new version of this commit will be simultaneously uploaded, which both fixes the bad behavior and allows us to have exactly 1 CL per branch that fixes the issue.

Bug: 118694079

Change-Id: Ic38a2ca2ddede7f0929779b0f2292b7823c11e87
Merged-In: Idab3ecc246fe9344a09e6907a0ba39f8ea6506f9