Deprecate Duplicate SignalStrength Methods
This change cleans up access and deprecates method
on the SignalStrength class that are no longer needed
due to the support of getCellSignalStrengths(). These
methods are generally technology-specific getters. In
addition, this tags non-public methods that were also
of this type to restrict their access to apps
targeting P or earlier via the @UnsupportedAppUsage
annotation.

In addition, deprecate the semantically incorrect
isGsm().

Lastly, deprecate the NotifierBundle() methods that
rely heavily on implementation details that changed
in this release, and which should also be unneeded due
to getCellSignalStrengths() giving access to the
internal storage classes of SignalStrength.

Bug: 122845432
Test: make update-api && make offline-sdk-docs && make
      (no functional change).
Change-Id: Iec794aee337099dee91d0820f7269fda433e8b01