Make it clear IMMS#updateSystemUiLocked() is always called with mCurToken
When InputMethodManagerService#updateSystemUiLocked() is called from
setImeWindowStatus(), the first parameter is guaranteed to be the same
as mCurToken, because calledWithValidTokenLocked() returns false for
null token.

This CL just clarifies it for the next mechanical refactoring.

This CL itself is also a mechanical refactoring.  Hence there should
be no behavior change.

Bug: 34851776
Test: atest CtsInputMethodTestCases CtsInputMethodServiceHostTestCases
Change-Id: I086dc0381078fc292db8132bc64d920ef7646c1c