Coerce Unreported BER from ASU=99 to UNAVAILABLE
Modems inconsistently use 99 or UNAVAILABLE to report
that BER measurements are unavailable; this causes the
new strict logic in SignalStrength to conclude that a
SignalStrength type has been reported when in fact it's
just the modem choosing the ASU out-of-range value
instead of the Android "unreported" value. We can't
force older modems to change, so this patch simply
coerces 99's to INT_MAX. It means that the ASU 99 will
never be returned on older HALs, which should have
no practical implication.

Bug: 123088652
Test: cts - atest SignalStrengthTest
Change-Id: I4c8610356a81e2190e475d789138b2d2c344fa36