Fix `break` location for WM Service message handler
The message handler for the WM Service is implemented in a switch
statement. For each case the `break` command is located in different
locations, making it harder to debug and easier to make mistakes.

Move the break conditions to the same location in all messages.

Test: Check the conditions of the `handleMessage` method. All messages
should end with a `break` command, within their body

Change-Id: Ibd225547d46ae1b092c071824570e5dbe01a8dca