Fix hidden usage sConstructorMap
In LayoutInflater, sConstructorMap was marked with @UnsupportedAppUsage.

After doing some research, the main reason for accessing this appears to
be related to classloaders (fixed in CL=839342, method=verifyClassLoader).
As this is fixed and creating a public API for a cache doesn't seem
apropriate, dark grey listing it seems the best solution.

Bug: 123769490
Test: android.view.cts.LayoutInflaterTest
Change-Id: I104ce17bb4083f7049bbb512090b31f8360cb23b