Kill app zygote when package is removed.
Since the app zygote is kept around for a few seconds after the last
isolated service using it goes away, it was possible to install a new
version of an app and launch a service before that timeout had expired -
causing the service to use the app zygote which loaded the old code. To
prevent this, clean up the app zygote as part of package process
cleanup.

Bug: 124401845
Test: manual
Change-Id: I2d2ef28280a9631275569e419283e160ca9e1105