Fixed incorrect data network type when device is on IWLAN
When the device is configured as AP-assisted mode, the cellular
network service should reports the cellular data RAT instead of
IWLAN. However, before we clean up all usages of
ServiceState.getDataNetworkType(), we still need to simulate the
old behavior, which is reporting IWLAN as the data network type.
Once all places are cleaned up, we can remove this temporary
solution.

Test: Manual
Bug: 124994459
Change-Id: I37256517af3bcf7330cefe55a33d6f09af7e1353