[Magnifier-78] Fix shadow on first draw
The CL fixes the light source position when the magnifier is first
drawn. Before this CL, the source position was set only after the
magnifier was drawn for the first time, causing an incorrect shadow for
the first draw. This was noticeable if the magnifier was trigerred and
not moved (the shadow was updated and started looking fine when the
magnifier was moved and a new draw happened).

Bug: none
Test: manual testing
Change-Id: Iabdd34430a5def3608197294cf694c5045a7d8d3