Notify the WMS when the display content is ready.
We only add wallpaper to the display that can support the display
system decoration attribute. However, when receiving a display
add callback does not mean that the display content has already
been created, we could not add wallpaper to the display in this
case.
To prevent above condition, we can notify the wallpaper that the
content is ready to add wallpaper instead of listening a display
was added.

Bug: 125007644
Bug: 124073468
Test: atest ActivityManagerMultiDisplayTests#testWallpaperShowOnSecondaryDisplays

Change-Id: I67062ce62cdf185884ab77a5741e5374e5987ef6