WM: Don't skip on parent changed when creating TestAppWindowToken
This was causing a few tests to break since a surface control was not being created and the tests
expected a surface control.
- ZOrderingTests
- AppChangeTransitionTests
- AppWindowTokenAnimationTests

Reverting the change did not cause any new test failures and fixed ZOrderingTests. If the change does not
cause any issues, will follow up with a cl to clean up the special case for TestAppWindowTokens.

Bug: 127948929
Test: atest WmTests
Change-Id: I2f374283472f1020a2792ca83de3f0fadc92009f