Require the ANDROID_TZDATA_ROOT env var
The ANDROID_TZDATA_ROOT serves a similar purpose to the
ANDROID_RUNTIME_ROOT: it determines the location that can be used to
find time zone data files. Since the Android Core Library will soon
require it to be set it makes sense to check it alongside the
ANDROID_RUNTIME_ROOT variable.

Bug: 128422035
Test: build / boot
Change-Id: Ie4860f60760905a06a744bd48fd773d0a4e8cf6d