Revert "MediaHTTPConnection: move connection states into an inner class"
This reverts commit fdf2e992a29156e68748ffde94fb35c7900660d8.

Many of the fields that were moved are annotated @UnsupportedAppUsage,
so the CL would have had undesirable app compat impact. Further,
because investigation has revealed that lock contention *is* possible,
we need to always acquire the lock anyway so there is no longer a
benefit in keeping all of the mutable state in a single field that
can be atomically set to null.

Test: Treehugger
Change-Id: I202c5653cb086d99228491e161a159bad640105a