OMS: always update overlays.xml when overlay is removed
The OverlayChangeListener in the overlay manger service is responsible
for both writing /data/system/overlays.xml and starting the flow to
notify affected processes. Because of this, the listener needs to know
when an overlay is removed regardless of wheather the overlay is enabled
or not, or overlays.xml will contain stale information.

Unconditionally notify the OverlayChangeListener when an overlay is
removed.

Test: atest OverlayHostTests
Bug: 129368140
Change-Id: I1f5aeb285154a5d90ff81161825a95d9325a479e