Only reset layout seq when moved to a new display.
In previous commit 494bfcebc02c77e41df999bbc1d6be9fd7144724 we
consolidated display info change and moving to new display into one
single method. Partner found that change may lead to timeout on turning
on the screen (needs more than 1s to turn on screen). Thus restrict the
layout sequence update to only moving to a new display, which is
effectively the same to it before the consolidation in the previous
commit.

Bug: 128812757
Test: go/wm-smoke. Partner verified this fix too.
atest WmTests:WindowStateTests
Change-Id: Ifd5d8e98468f5fdb4bf9ec0de14766b58b36085e