Do not hide caption for hiding navigation bar flag
Previously, if the app is in fullscreen and with system UI controlling
flags to hide/dim the navigation, the caption will go away. Any of the
following four flags will hide caption: FULLSCREEN, LOW_PROFILE,
IMMERSIVE and HIDE_NAVIGATION.

IMMERSIVE flag itself will do nothing without HIDE_NAVIATION and
FULLSCREEN. LOW_PROFILE should not apply to caption as the caption
cannot be dimmed. HIDE_NAVIGATION should not apply to caption. We should
only hide the caption when the FULLSCREEN flag is on.

Bug: 131099092
Test: Manual with latest Gmail app.
Test: go/wm-smoke
Change-Id: I4757c4a4ccbe1cc106b5b6f856f3e734bbd6cbb6