LayoutInflater: Special case well-known View constructors
This avoids loading constructors with reflection for well-known View classes.

An average of 1300 app startups with and without this change shows the special
casing improves app startup time by about 2.5ms.

The classes listed here are taken from examining several app traces as well as
the similar list in AppCompatViewInflater.

Bug: 131421854
Change-Id: I676a50eec50b86fa0b385add4bc092a657d8e8bb