Revert "MemoryIntArray: don't double close."
The reverted change causes a regression where we can get an IllegalStateException
during finalization as we are adopting the native fd in a ParcelFileDescriptor
which takes ownership of the fd. However, the order of finalization is undefined
and if the ParcelFileDescriptor is finalized before the MemoryIntArray we would
get an exception when running the finalization of the latter.

This reverts commit 52ad77e8d847fee58f7292485aa637cc6b2f1a1d.

bug:124056170