Fix RTL in QS Header Info
When there was only one view to display, we relay on
FrameLayout#onLayout. Setting the layout_gravity to center_vertical
(which is unneeded here) was removing the "start" by default.

Lays out Views from start to end keeping track of offset. As the Views
were given their final size in onMeasure, laying out is easy.

Fixes: 133221093
Test: manual using Hebrew and English
Change-Id: Id57d951bbbacbff7745869fe5b288d620a3c25f1