Ensure that the Application is constructed before any Service.
This fixes dependency injection issues and brings Service construction
in line with the way that Activities are created. Specifically, this
ensures that there is always a reference to a Context available at
the time a Service is constructed, such that the Context can then be
injected into the constructor of the Service.

Bug: 140868981
Test: atest SystemUITests
Change-Id: If85e4106fef8c2fd12f7ca9e8b7fd686af62e29b
(cherry picked from commit 25262cafc3cb9998ed09db8e0e8b1e9dd58d5e77)