Fix device not locking properly on power press
A recent change to start listening for fingerprint auth sooner when the
device is going to sleep caused an issue with the device not locking
properly in some cases when the power/lock button is pressed. This
commit fixes that issue and updates the unit test to check for the
offending method call.

Test: atest KeyguardViewMediatorTest
Test: Unlock phone, press power, verify device is always locked
Test: Press power to lock, quickly touch sensor, verify device unlocks

Change-Id: I11b388b16184a6903ba7f4f5cf977e91d79efd90