Support privileged apps installed in APEX.
If an APEX contains a priv-app/ directory, and the APEX package itself
is located in a partition for which privileged apps are allowed, then
the APKs under said directory will be included in the initial scan and
made available to the system.

Bug: 138429615
Test: Move PermissionController to com.android.permission APEX. Run
CtsPermissionTestCases and verify that the 16 failures out of 264 tests,
which seem unrelated to PermissionController, happen also on an
unpatched build on head.
Change-Id: If809db5d0a061c1a55aeb7f830c183fa822adca5