Replace framework-annotation-proc java lib with framework-all
The purpose of the java lib framework-annotation-proc is to run some
annotation processesors on source files in framework-all-sources.
However, running an annotation processor requires the compilation
of the entire source files; in fact, the output of the annotation
processors are created as byproduct of the compilation. This is a
duplicated compliation because we already are compiling the source files
for the framework-all java lib. To reduce the duplication, replacing
the references to framework-annotation-proc with framework-all, and
removing framework-annotation-proc.

Test: m
Merged-In: I449488cba235211374e2825df97aae84395e60b6
(cherry picked from commit 3b9f618a54bb30edaedf916e3ccab4b16d24d771)
Change-Id: I449488cba235211374e2825df97aae84395e60b6