Add package name to DDM APNM packet
DDM APNM currently only reports the application/process name. For
services that need to know the underlying package name, those
services need to manually resolve the process's UID and map it to
the package name via the shell - a slow and error-prone process.

This change adds the package name to the APNM packet so users of
newer versions of DDMLib can directly get the package name.

Test: manual

Change-Id: I92517d5bb641cc263c0d58db667b00e67fc5e58b