Allow Windows in Display Overlays
Instead of display overlays being a SurfaceControl that can only have
child surfaces, it now is a WindowContainer that can have windows.
Specifically, windows with the type TYPE_ACCESSIBILITY_MAGNIFICATION_OVERLAY
will be placed in the display overlays layer

This is the first step to allow window magnification to use real windows
on the display.

Test: go/wm-smoke
Bug: 136250281

Change-Id: Idad98db7a85f68ba49557c0efb83b9a1be089609