Tell StorageSessionController when a user is stopping.
When a user is stopping, all processes running under that user will be
killed, including the MediaProvider process that hosts the FUSE
filesystem. This means that the binding to the MediaProvider will fail;
to prevent us from trying to rebind needlessly, let the
StorageSessionController know that a user is about to be stopped.

Bug: 137890172
Test: test started and stopping a user; no attempts to rebind.
Change-Id: Ie2810464f20f13a9700ead506c82731f603ab374