Add MANAGE_APP_OVERLAY_PERMISSION intent action
In order to support system use-cases, such as when the user wants do
revoke SYSTEM_ALERT_WINDOW (SAW) permission via notification, we are
introducing a new intent, MANAGE_APP_OVERLAY_PERMISSION (notice the
"APP_"), that directs the user to the app-specific screen to toggle SAW
permission.

The intent can only be started if the caller has INTERNAL_SYSTEM_WINDOW
permission, which sysui and the system have. This is verified in CTS.

In this change we also update usage in NotificationGutsManager and
AlertWindowNotification.

Test: Builds
Test: atest TestManageOverlayPermissionIntents
      (cts in one of the topic CLs)
Test: Let an app draw on top of other apps, via its notification go to
      settings, verify that the screen for controlling such capability
      for the app is dispayed.
Bug: 145198756
Bug: 135920175

Change-Id: Iba29e1eaee729268a4dec9561946f6ed74b36619