Hide framework-minus-apex
framework-minus-apex is the jar implementing APIs that the non-updatable
part of the framework provides. It shouldn't be used by modules that are
unbundled. Currently, several modules are using it directly because they
haven't finished clearing their private API usages. Until the work is
finished, allow framework-minus-apex only for those known modules.

Bug: 147128803
Test: m
Change-Id: I09204b7d600a4d0cc1f6d248df6c813bfb8bae01