Recover from FUSE session starting failures
Vold calls into StorageManagerService which in turn calls into the
FUSE daemon to start the session. If the system_server is very busy
e.g right after an OTA, the session may not start quickly enough and
we will timeout, causing vold to fail the mount.

Previously, the only solution would be for the user to reboot.

Now, we naively retry every 10s.

Bug: 147610762
Test: Causing an artificial timeout when starting a session for the
firs time, verified we recover after a minute

Change-Id: Ifdb2384696e85b6774a12394798fca8eb75a584e