Re-seal the sessions on reboot.
Previously we were validating them as well. This is unnecessary as
validation will happen on commit/transfer calls later.

This was introduced in 2eb73d9b087810810cfe9b03455a3686cea81bd4 when we
were using mSealed flag for both sealing and validation. It's not the
case anymore and we can safely deffer validation to commit/transfer.

Test: atest PackageManagerShellCommandTest SplitTests StagedRollbackTest CtsAtomicInstallTestCases PackageInstallerSessionTest InstallSessionTransferTest
Bug: b/136132412
Change-Id: Iedba991ff057bcfc4aeb2ba0d450e59bb4f035d8