Add BubbleCoordinator
When FeatureFlags.isNewNotifPipelineRenderingEnabled, the
BubbleController will use the NotifPipeline as its source of
notifications instead of the NotificationEntryManager. Note: in this new
pipeline, the GroupManager is not longer necessary; instead, the
ShadeListBuilder handles grouping. The BubbleCoorinator handles
filtering and removal intercepts (see BubbleCoordinator documentation
for more information).

Note: In BubbleController.handleSummaryDismissalInterception
we now intercept auto-generated summaries so the children
of the group aren't dismissed too early in the new notif pipeline. These
summaries will still be removed by NEM via the maybeCancelSummary
method. In the new pipelipe, the summaries will be pruned in
ShadeListBuilder and removed when system server gets back to us to
remove it.

This CL also includes a fix for a bug that occurs when a user-dismissed
summary is associated with bubbles that are hidden from the shade AND
regular notifications in the shade. Previously, if the summary was
dismissed, the summary would be intercepted and the children that
weren't hidden from the shade would show in the shade (ungrouped)
even though they should be dismissed. The fix is to manually dismiss
these children in BubbleController in handleSummaryDismissalInterception.

Test: atest NotifCollectionTest
Test: atest NewNotifPipelineBubbleControllerTest
Test: atest BubbleControllerTest
Test: atest SystemUITests
Change-Id: Id1639f2a5997103b2fd14971433b764c3b1b0112