Create KeyguardViewController that extracts view-related logic.
Prior to this change, KeyguardViewMediator used
StatusBarKeyguardViewManager for making view-related changes.
This meant that KeyguardViewMediator could only be used with
a specific Keyguard view that is mounted to the StatusBar.
This change makes KeyguardViewMediator compatible with any
Keyguard view no matter where it is mounted, as long as it
implements the methods defined in the KeyguardViewController interface.

For AAOS, This refactorization allows us to implements
our own Keyguard View and its Manager as the next step.

Bug: 150467593
Test: Build and Manual Test -- Android builds successfully on Hawk and Pixel 2 and
interacting with Keyguard functionalities works.

Change-Id: Idda93514eac030f1d92523aebc9444a4d55b21d2