Explicitly exclude IME from Task Snapshots
In Q we built a mechanism for excluding layers from snapshots for
excluding the IME from TaskSnapshots, but somehow we never
ended up using it. I guess instead we were relying on
reparenting or timing. Whatever we were relying on is somehow no longer
working. Seems best to fix it with explicit exclusion so we don't
have some hidden dependency.

Bug: 148953305
Test: Open IME. Enter recents.
Change-Id: Iea2541ba3a1943e9137d6681b2a6f876bb63a392