Remove WebView fallback logic migration code.
The WebView fallback mechanism was removed in Q and replaced with code
to migrate the state of upgraded devices to the new configuration. This
is no longer necessary; remove the logic to simplify the code.

The notion of a WebView provider being designated as the fallback is
kept, as this is still used by the code which checks if there is no
valid provider at all at boot time and tries to fix the situation by
re-enabling the fallback provider for all users - this was added for
b/132815849 and should suffice for any devices which haven't executed
the fallback migration code.

Bug: 129470358
Test: atest WebViewUpdateServiceTest
Change-Id: I4de777b650fb793903f6df74a7ff639190d8b0a9