Fixes NPE and adds @Nullable to PackageSetting.pkg
This change addresses an NPE encountered when a package setting refers
to an unavailable package and is thus null. It adds a @Nullable
annotation to the field to hopefully catch more of this type of error
going forward.

Fixes: 153007501
Test: atest AppEnumerationTests PackageManagerTest
Change-Id: I0ffe1f488d69d45893c96c755e48e19de43fcdf2
(cherry picked from commit 11021bef9395d077e2d04dc2c300c480fcbd8195)