Dispatch bind pipeline at end of UI thread handler
To prevent constantly starting and stopping async tasks in the inflation
pipeline when one change causes multiple rebind requests, we push the
actual starting of the bind pipeline onto the UI thread to be handled
after all of the current message is finished. Now, multiple rebind
requests simply put ONE message on the NotifBindPipeline handler to
dispatch the pipeline.

Bug: 150719232
Test: atest SystemUITests
Change-Id: I92ee530ad5513f3ca015e82c4935e9e5693365b4