Consolidate TaskDisplayArea#addStack and TaskDisplayArea#addChild
When closing an app by swiping up, launcher runs a remote animation
on the home task and the closing app task. If the most recent running
task (via ATMS#getFilteredTasks) does not match the remote animation
 target task then the animation is not run and we see a flicker.

When dismissing split screen, the recency of the primary split screen
task is not updated. Fix this by eliminating two different code paths
to add a child to TaskDisplayArea and ensure we update
TaskDisplayArea#mPreferredTopFocusableStack.

Test: swipe up to dismiss app after exiting split-screen
Test: go/wm-smoke
Fixes: 153583420

Change-Id: I592199d95f3f2ae56ad5ff638612cb63a3549602