Update GlobalCellId during Un-parceling
The Global Cell-ID is constructed from other
fields in each CellIdentity.

Currently, it is not stored during parceling.
Thus, after a CellId is parceled and unparceled
this value is being lost.

Do a quick fix to update the global Cell ID during
unparceling. A longer-term fix is to resolve the
layering violations that create the situation where
the global Cell ID is defined in the base class
but manipulated in the derived classes.

Fixes a broken unit test.

Bug: 154955167
Test: atest ServiceStateTrackerTest#testRatRatchet
Change-Id: I5c58322ebcec8886405e25e09754c7f154187efe