Add latching threshold for ThresholdSensorImpl
This allows ThresholdSensorImpl to account for noise in their sensor
output. They must go below one threshold to trigger a "below" output,
and then must go above a different threshold to switch back to an
"above" state.

This new second, latching threshold is optional and defaults to the
existing threshold if not set.

For the ProximitySensor, we introduce proximity_sensor_threshold_latch
and proximity_sensor_secondary_threshold_latch in config.xml.

Fixes: 147026387
Test: atest SystemUITests && manual
Change-Id: Ie0b7b357a48fd27424d12b890cda9e00073e1eb3