Sends the correct error code
In ag/11269881 we cleaned up the parent session from the
child intent status listener. While that's ok in concept, we've already
translated the error to the public value at that point and so use the
wrong values to determine whether the session was a failure. In this
change, we fail and clean up the parent from the original failure and
just do a bit more filtering in the child intent status listener to
ensure we don't send duplicate broadcasts.

Test: atest AtomicInstallTest StagedInstallTest
Fixes: 155509347
Bug: 144567152

Change-Id: I55c8cf144413c48b2aaa48c1ae2d5f22ebdcc13a