Revert ResolverActivity changes associated with browseables.
Reverts changes 72a41b4, 1a6c2741, 87919b7 and others. These changes
don't revert cleanly because of several refactorings layered on top
of the original changes.

The main objective of these change is to get rid of mUseLayoutForBrowseable
and associated codepaths as we treat choosing between browseables the
same as other choices.

Bug: 157460946
Test: manual
Test: atest ResolverActivityTest
Test: atest ChooserActivityTest

Change-Id: Ibe9f2289289f7f5da3986e6892a2ee4ff65765a0