Allow wallpaper to be visible if wallpaper target exists.
When entering PIP task, the obscured condition will be changed after the
fullscreen task reparent to PIP task, and since the PIP animation is
moved out, the opening animation of enter NexusLauncher will not synced
with SystemUI, so wallpaper could become invisible during we moving a
fullscreen task to PIP task.
So a simple fix idea is that if the wallpaper target is visible, then
wallpaper should also be visible.
And review the findWallpaperTarget method, we find the check conditions
have been duplicated with isWallpaperVisible, so we can simplify the method
isWallpaperVisible by checking whether the wallpaper target exists.

Fixes: 156014070
Test: atest WallpaperControllerTests
Change-Id: Ib02f1e6ce10d4c412d049aa84541f5a5323fd455