Switch framework-statsd to use java_sdk_library
The names of the individual modules do not quite follow the pattern
that java_sdk_library uses so this temporarily sets the following:
    naming_scheme: "frameworks-modules"

That causes java_sdk_library to use a naming scheme that matches the
one used by the individual modules of this. It will be cleaned up
later.

Test: m update-api
Bug: 155164730
Change-Id: Iadb4020a8b97c7f1eed8ee6752d91e827404441e