Disable shared library support for framework modules
Defaults shared library to false as the majority of framework modules
should not be usable as a shared library. The exception is ike but it
does not currently use the defaults and will override this setting
when it does.

(cherry picked from fad2181b36c3252dc73547798bfb5c39f4c7253f)

Bug: 156723295
Bug: 155164730
Test: m droid and flash
Change-Id: I3e78994a9855110b02c3d71944ff38b986208e9a
Merged-In: I3e78994a9855110b02c3d71944ff38b986208e9a