PiP: Potentially hide menu immediately after it appears.
PiP currently shows and hides depending on HOVER_ENTER/HOVER_EXIT
events. However, it is possible that user does ENTER/EXIT so quickly
(e.g. fling the cursor across the PIP screen from one corner to the
other), that by the time HOVER_EXIT has happened, the menu activity
hastn' finished showing up yet. If that's the case, #hideMenu would be
called too early, which ends up not hiding anything.

This saves a temporary check such that when the menu appears, if there
is already a signal to hide the menu, then we choose to hide the menu
immediately after.

Bug: 122982825
Test: Fling the cursor from one edge to the other, no longer see the PIP
menu stuck showing.

Change-Id: I946cd2614c0270ffacbaf7d7e6942891095c03ba