Correctly initialize recognition events upon abort
Aborting a recognition as result of capture being enabled for devices
not supporting concurrent capture is the only case of creating those
events that is *not* conversion of the HAL events.

Unlike their HIDL counterparts, object and array fields of AIDL
parcelables are NOT default initialized and so are not considered
valid by our conventions. This change make sure those events are valid
by default initializing all the fields and setting the model type
correctly.

Fixes: 157496890
Test: Manual verification of existing sound trigger cases against
      regression.
Change-Id: I4043997d6d75f1fcb8525dcbd051b13ae2bf1d1b