Only apply server visible to split screen
When the app windows switching, the control target can temporarily
updated to the remote control target and the insets will become server
controlled in that case. However, we don't want to apply the server
visibility for that case as it's not a real insets update.

This patch changed the insets visibility calculation, to consider server
visibility only if the display is in multi window mode and the insets is
not client controlled.

Test: atest InsetsSourceProviderTest
Test: go/wm-smoke
Test: See reproduce steps in the bug.
Bug: 158447971
Change-Id: I0525813459b367d23d8300d122c72a2b49bac967