UserLifecycleTests - give proper errors on fail
More 'attestation' messages are added, which
provide logcat messages in the case of failed tests
while telling the benchmarkrunner to mark the test
as having failed (and return a value of -10).

Tests no longer throw checked exceptions; we catch
all such exceptions and log them properly.
The exception is RemoteException, which we have
allowed to throw for now but may handle in a future cl.

Test: atest UserLifecycleTests
Bug: 159650471
Change-Id: I4ca9d7186d195388f928f6a6037c6d4ccd79a0ea