Stop keeping freeform activities visible.
This logic was introduced when there could be multiple tasks in the
freeform stack so it was necessary to keep freeform tasks visible even
when they are not in the front. However right now stack and task are the
same so activities became the children of the original stack, and we
don't have the need to keep activities behind the top activity always
visible.

Bug: 161366153
Bug: 177403437
Bug: 179247198
Test: atest ActivityVisibilityTests#testConvert* on freeform first
      devices.
Change-Id: Ibeeba5ca8b6b60583eec9cbee9b7209ee17b1037
Merged-In: Ibeeba5ca8b6b60583eec9cbee9b7209ee17b1037
(cherry picked from commit 1de74ec6a004752dab58487fbaac2e5689e15862)