Unset activity type for process level configuration
The process level configuration was updated and was scheduled
to the client when home activity started, in which as home
activity type. This may cause inconsistency when a standard
type activity is started in the same home process.

Bug: 162030251
Test: atest WindowProcessControllerTests

Change-Id: I1b3b9a1a9cd164f784142fcacb365240d3522221
(cherry picked from commit 642c74d6d9c14f155dd6876bc8417dbcedf25dfb)