Fix NPE when starting activity while no focused task
NexusLauncher was not started when FallbackHome finished itself
because activities was prevented to be resumed while display is
sleeping. So, NPE was thrown while starting an activity because
there was no focused task in the display since FallbackHome
activity was removed from the task and all tasks were also removed.

The activities should still be able to be resumed when calling
resumeTop methods explicitly (even when display is sleeping).

For the original issue of bubble activities, the activities in
secondary display can be prevented from starting by evaluating
keyguard-going-away state for default display only.

Bug: 164572568
Bug: 160338354
Test: atest ActivityStackTests

Change-Id: Ica58b7e8c93e7aa688b335d1ecf7464bc8450e30