HexDumpTest: Fix assertion.
Commit 80001af23c7560c59d0a5581e14cf2f59e6d30d8 added test coverage,
including a helper function assertThrows(). The implementation of that
method was buggy because it only failed the test if it was throwing the
wrong kind of RuntimeException, instead of also when the code under test
wasn't throwing an exception at all.

This CL fixes the helper function. The test still passes (none of the
code under test was relying on the incorrect assertion).

Test: atest FrameworksCoreTests:com.android.internal.util.HexDumpTest

Change-Id: I8fe3e70ada8b880a03e124cb48e204e7bb880e57