AudioService: more communication mode fix
Do not reset communication mode when an app is not playing or
capturing audio if the app is not the audio mode owner.
In case of dual call situation, a VoIP call can be placed on hold and it
is legit for the app to remain in the mode owner stack but pause
capture and playback.

Bug: 161564483
Test: VoIP call interrupted by Cell call.
Change-Id: Iac5bee9b606e3663b63a4c1c58f9c7a880ab0311
Merged-In: Iac5bee9b606e3663b63a4c1c58f9c7a880ab0311
(cherry picked from commit dfc3bc757fe7a84aefba78d8ad3df0c2ad723f2a)