Set default value of A11y magnification capabilities
We set magnification capabilities all by default, but if the user
is using full-screen magnification before migration, we will set it to
full-screen.If the above situation happens, we show a prompt dialog to
inform theuser the new feature when using full-screen magnification
first time.

Note the default value will be applied if the user erases all datas.

Test: 1. Enable/Disalbe full-screen magnification and observe
         the values after flashing.
      2. Erase all datas and observes the settings values.
      atest SettingsBackupTest
Bug: 162525393
Change-Id: I3dc749c2414f3aa9162c0f2a31b512c5e6a3ae80