Deprecate PackageParser main class and add OWNERS
This wasn't deprecated in the original refactor because there was no
publicly exposed alternative that encapsulated all the functionality.
There are also some subclasses which are still used as they haven't
been copied to the parsing subpackage yet.

But this has lead to contributors making changes to PackageParser, not
realizing that the class is unused. So it's worth marking it deprecated
just so it's obvious to anyone editing the code.

The changes made since last release seem mostly unconsequential, so
they will not be reverted, but no additional new code should be added.

Also protects PackageParser and core parsing with OWNERS.

Test: none, docs change

Change-Id: I9d80132503b3a5a3384abbe6b0b1fbb752029710