Remove the unneeded icon spacing in the InboxStyle.
Because the LargeIcon is was moved up a 'line' to
the top of the notification, the only lines which
can intersect it are the header and title lines;
the first line of text from the inbox style no
longer can.  This also sort of depends on
ag/13018219 which fixed the vertical spacing
enough to make this look good.

Bug: 163626038
Test: manual
Change-Id: Ic9ed82eb3c0e9c7e619917351c3461ffe1713857