Run error callback when BiometricPrompt dismissed early
Currently, if BiometricPrompt is dismissed quickly (e.g. by tapping on
the top half of the screen to dismiss it while it's being shown) the
ongoing authentication session may not be canceled and the
onAuthenticationError callback may not be run. This commit fixes the
issue by ensuring that a dismissal reason is sent to callback if the
prompt is shown while in the "pending dismissal" state.

Test: atest com.android.systemui.biometrics
Test: Manual:
1. Using a custom test app, trigger BiometricPrompt to be shown
2. While the prompt is being shown, repeatedly tap top half of screen
3. Ensure the prompt is dismissed and onAuthenticationError is called

Fixes: 173417042
Change-Id: I240efba1b008681534380c5e4a10bff973f258a6