Ensure notifying when only caption insets changed
There's a bug where we dispatch the insets state only compare to the
previous dispatched state excluding the caption insets. Make the insets
included to make sure the caption insets change can be dispatched if
only the caption insets changed.

Another issue is, when we dismiss the caption, it happens after the
window lay out traversal but before the post layout. It is necessary to
notify the insets change because there's no server insets change in the
post layout process and the app won't be able to receive the insets
change.

Test: atest InsetsControllerTest
Test: see reproduce steps in b/174718377
Test: Turn an app into freeform mode, the insets will be applied
      immidiately
Test: go/wm-smoke
Bug: 174718377
Change-Id: I5c40c406fcec40c9fe87ef6fa1b54e1a2592e3bc