Revert "Revert "Fix permission on TM#isDataEnabled""
This reverts commit 45ed2f6ca8c213dc6d3cd3e94d6e2de7dad6c61f.

Bug: 176163590
Bug: 148500541
Test: manual smoke test on low-end device and aosp_crosshatch builds.
Reason for revert: Reapplying with a fix for the exceptions.

The exceptions were caused by an undocumented behavior difference in the APIs that just blindly catch runtime exceptions. The cause of the runtime exceptions is that people are calling isDataEnabled() before telephony is running; however, until a "better" fix is available, this API will apply the existing behavior of just catching the runtime exceptions (which is terrible, but it's the plan of record in the code today).

Change-Id: I9fbe7126ee7039dfb0b383f83017458a3525db66