Add REBOOT_READY broadcast
Adds a new REBOOT_READY intent
which may be sent from the system server,
indicating the device's reboot-readiness state.

This broadcast will be sent with a boolean extra,
which will provide a positive or negative signal
of whether the device is ready to reboot. This
broadcast may only be received if the REBOOT
permission is held.

Test: atest CtsRebootReadinessTestCases
Bug: 161353402

Change-Id: I70fa1c35c844189af86c519749144732ceceaaac