Always use the latest CLC for system server
Overwrite the class loader context for system server (instead of merging
it). We expect system server jars to change contexts only in between OTAs
and otherwise to be stable. So it is ok to take the latest context all the
time instead of trying to merge, which will lead to the context being
marked as variable and thus making dexopt sub-optimal.

Test: PackageDexUsageTests DexManagerTests
Bug: 179512101
Change-Id: Iaf8368e4ededa6eab074dfe9589fbdaac24491d1