Add cpu proc file read throttling enabling/disabling
To properly attribute the new MeasureEnergy data for CPU, the cpu kernel
uid file reads needs to occur everytime the CPU MeasuredEnergy is read.
If throttling is necessary, it will need to be handled by
BatteryExternalStatsWorker for both the cpu kernel file reads and the
MeasuredEnergy read to make sure the data is in sync.

Bug: 180079165
Test: atest FrameworksCoreTests:com.android.internal.os.KernelCpuUidUserSysTimeReaderTest
Change-Id: I90ebad48d780872a611605689e0d25678da66f27