Check and create data and obb dirs after starting zygote process.
When data or obb is not cached, it may slow down app starting speed as
we wait the directory is paged before starting zygote.
To improve this, we postpone the dir check after starting zygote process
in ProcessList, so the directory paging and creation can be done in
background without blocking app starting.

As it's possible that dir is not created when zygote needs it, we also a
busy waiting check in zygote to ensure data and obb dirs are created before
we do our work.

Bug: 177248242
Test: Able to boot device and start app after enabling app storage data isolation flag.
Change-Id: I892b7a1a4f8c38331546f9129b8fb43333093aec