Wait for unlock to start jobs.
We don't tell apps that a user is unlocked until the user is fully
unlocked (not just unlocking) so some apps may think that storage is
still encrypted even when it's not (during user unlocking). Wait until
the user is fully unlocked to run any jobs for a user to avoid potential
problems.

Bug: 162076852
Test: atest frameworks/base/services/tests/servicestests/src/com/android/server/job
Test: atest frameworks/base/services/tests/mockingservicestests/src/com/android/server/job
Test: atest CtsJobSchedulerTestCases
Change-Id: I4c7e2e2e4f86162c59685295417e02eb554c1cb0