Do not force endDocument for domain verify settings
Removes the endDocument call so the serializer can be used as a child
of a larger group. Instead calls flush and assumes all usages will
close tags appropriately.

This usage is incorrect regardless, but endDocument itself is broken
and does not behave the same depending on which XmlSerialier
implementation is being used. Which is why this bug never appeared
during testing/development.

Bug: 180838875
Bug: 181813200

Test: manual, debug linked Bug
Test: atest DomainVerificationPersistenceTest

Change-Id: I50e27817e526d2cbcc0efbcc3ffdfffa53886f0f