Set source package when copying from share sheet.
This uses a new system API to attribute the source package when copying
to clipboard from the share sheet.

This is needed because in S users will see a toast message saying
"<app> pasted from <source>" when an app accesses clipboard. Without
this change, any copy originating from the share sheet results in
"<app> pasted from Android System" rather than the app that is the
actual source of the data.

Bug: 180577866
Test: Manual, share via share sheet and paste in another app;
observe "<app> pasted from <source app>"

Change-Id: Id1c0afcf6e775868febc7c8ab5b44dbacccacfa2