WindowManager: Handle main window size change transaction with BLAST.
Rather than push and defer the pending transaction we can use the
new applyWithNextDraw primitive to make sure the transaction applies
with the app redraw. We need to be a little careful. Since
mainWindowSizeTransaction is consumed during layout, we need
to move the check for pending draw handlers to the end of relayout.

Test: Existing tests pass
Bug: 168505645
Change-Id: Iae7906ee6def94ee3adc71b5f5ba022157379c73