Fix notification colors on theme change.
Causes:
* The theme is implemented with RROs
* RRO state is stored in the ApplicationInfo's overlayPaths
* Notification.extras contains the ApplicationInfo from when the notification was posted

The fix here is to fetch and inject the latest ApplicationInfo (using a method which has a correctly-invalidated cache) right before recovering the builder from the notification to bind content.
Also, because this codepath is operating on the background thread, there are no SystemUI jank risks during the overlay change.

Fixes: 182662365
Test: change theme; notifications change color
Change-Id: Ibdfaf811155d64e0ecb7c70075e3ba4429c6d920