Only allow local sessions to be resumable
It is possible for an app to be resumable locally, but also post
notifications when playing remotely that are not resumable long-term.
This change makes it so that only local sessions will be used for resume
players, and remote sessions will be removed when the underlying notification
is removed.

Bug: 185681682
Test: manual - start playback on Spotify web, observe notification on
phone, then close app on phone and observe no resume player
Test: atest MediaDataManagerTest MediaResumeListenerTest

Change-Id: I0a6ff2bab234a33298f84fd6d99b7d537d3d5e34