Fix black screen after moving top activity to another display
The screen was black after moving the top activity to another display.
The next top activity did not be resumed and was still invisible. The
regression was from b133671, which incidentally reverse the condition.

Since the case of removing a display is already handled by f7e4c95,
#postReparent is not needed to call again in #reparent (when display
is removing).

Fixing the issue forward by resuming the focusable task in
#moveRootTaskToTaskDisplayArea.

Bug: 192616155
Test: atest ActivityTaskManagerServiceTests
Change-Id: Id0971ad8a1c692fdb5376e2c077faa6f00dd61ac