Correct StartingWindowInfo for non-top activity.
When finishing a top activity and goes to resume the below activity, if
the resumed activity was gone and need to recreate, this should be treat
as warm launch and there should show an empty style splash screen on it.
Previous there do create a splash screen window but the parameters are
wrong because we are trying to create the StartingWIndowInfo from the
top activity.

Fixes: 193040222
Test: atest ActivityRecordTests
Test:
1. enable "Dont keep Activities"
2. Launch twitter, open link.
3. Click back key and there should show empty splash screen.

Change-Id: I33ac43ef6ca79c7422a818e468a7af32c7894eeb