Fix Keyguard exit remote animation not invoke finish callback
When Shell transition is enabled, some launcher tests failed because the
keyguard exit remote animation never invoke the finish callback.

The expect flow is:
1. Dismiss amount changed from 0 to 0.1+ to trigger the keyguard going
away transition.
2. Dismiss amount changed from 0.1+ to 0.3+ to finish the keyguard exit
remote animation.

However, in presubmit test, the dismiss amount may changed from 0.0+ to
1.0 directly. When it happened, there is only a trigger to start the
transition, but never trigger the change event to call the finish
callback, which blocked the Shell transition.

The idea is to also check the current dismiss amount when the keyguard
going away transition reach to the controller.

Bug: 183993924
Test: atest PromiseIconUiTest
Change-Id: I95951385130578f45df0837200cc7e88d89edd34