Allow moveActivityTaskToBack on aliased activities
When checking if the task root should be moved to the back following a
back button press, we need to ensure the behavior matches when
activities are launched via other activity-aliases for which it is a
target.

This means we need to check if the root activity matches the
realActivity (the resolved target activity of aliases) instead of the
base intent (which could include unresolved aliases).

Bug: 195362544
Test: manual with test app
Change-Id: Idb3ae6aba709c3dcfbe6cae01f46e4bafe7ab94f