Daggerize Plugin code.
Restructure the dependencies within all the plugin code
and make it suitable for injection.

There is still a lot of gnarly code in the plugin architecture,
and this is the first step towards cleaning it up.

This is also the first step towards being able to load plugins
(like the FlagReaderPlugin) _before_ the dagger graph is setup. With
this change, it is theoretically possible to directly construct
the plugin infrastructure, rather than relying on calls to
Dependency#get and other SystemUI idiosyncracies.

Bug: 194781951
Test: manual && atest SystemUITests
Change-Id: I04da3d12211d9f9d1a5c2c5cd27b6a2d81c3907e