Set uninitialized frame number to 0
Now that the frame number is a uint instead of an int it's clearer if the uninitailized value is 0. Frame numbers start from 1 so 0 doesn't collide with any valid frame numbers. It also means comparing frame numbers to check if the frame is newer always works even when comparing to the uninitialized frame number.

Test: N/A
Change-Id: I0901ffeaefbdd6378e59d8858a06de92b0104cd9