HDMICEC: Implement active tracking of CEC Network for Playback devices
Poll all the non-local logical addresses when the device is either turned on or when the CEC setting is enabled. After that the addresses are polled once every 1 minute.
Implement method inside HdmiCecLocalDevicePlayback that launches a DeviceDiscoveryAction with a HotplugDetectionAction inside its callback.
Add unit tests for DeviceDiscoveryAction and HotplugDetectionAction for HdmiCecLocalDevicePlayback.

Bug: 199057329
Test: make && atest CtsHdmiCecHostTestCases
Change-Id: I1133cdbced4a1e40e299a95bb9ced565c8059774