[RemoteInputView] Introduce RemoteInputViewController
Add a new Controller for RemoteInputView that will eventually house
all of the non-View code currently located inside of RemoteInputView.

This CL establishes the new Dagger configuration needed to instantiate
the Controller, as well as the lifecycle management by the parent
Controller (in this case, NotificationContentView, which doesn't have
a Controller of its own; once a Controller is introduced for that
class, the reference to the RemoteInputViewController will move there).

Bug: 193539698
Test: mp sysuig, receive some notifications, nothing breaks
Change-Id: Ib5397abe03569d2285764b69136ef44b903b1247