Fix missing notification footer.
A refactor to the DeviceProvisionedController changed the behavior of the addCallback methhod so that the methods of the supplied callback were no longer called.  The NSSL Controller depended on this behavior, so we used the default initialization state of the variable, not provisioned.  This hid the footer until a provision change.

Change-Id: Ie440b33dc344eb75bc026465606269850984d0a9
Fixes: 203744291
Test: manual