Allows running TaskFragment related hierarchy op in lock task mode
The placeholder activity was removed while device folded, but
the TaskFragment was failed to be removed because the hierarchy op
was ignored in lock task mode.

Most of the TaskFragment related hierarchy ops operate in the
same task and it won't change the task order. Therefore, it should
still works in lock task mode. However, deleting a TaskFragment
may cleanup all activities in the Task, which needed to be handled
separately.

Bug: 204945438
Test: atest TaskFragmentOrganizerControllerTest
Change-Id: If1d35ef81754b53c5f214672000cc0c066628a29