Check for the vibrate notification key on delayed vibrations
Notifications played with sound have a delayed vibration, in order to
match the vibration and sound effects. If the effects are cleared before
the vibration starts, or a new one takes its place, then the vibration
hould be cancelled because it no longer will be tracked by the
NotificationManagerService.mVibrateNotificationKey.

This can cause repeating vibrations, like incoming calls, to start and
continue vibrating even after the notification was handled, (if the user
opens the notification panel before it actually starts vibrating, for
example).

Fix: 202752953
Test: NotificationManagerServiceTest
Change-Id: If3ed4a741e78058a321691acebe735c735542f16