Include new api to exlude a shortcut from launcher
Include a new field in ShortcutInfo which serves as an indication of
whether a shortcut is exlucded from launcher. Shortcut marked as
excluded from launcher will not be included in the search result in
LauncherApps nor ShortcutManager. This generally means the shortcut
would not be displayed by a launcher app (e.g. Long-Press menu), while
remain visible in other surfaces such as assistant or
on-device-intelligence.

- setDynamicShortcuts/addDynamicShortcuts/pushDynamicShortcuts:
Shortcuts that are marked as hidden from launcher are ignored.

- updateShortcuts:
Similar to enabled/long-lived, developers cannot make shortcut hidden
from launcher by calling this api. An exception would be thrown when
updating a shortcut that is hidden from launcher.

- remove APIs:
Unchanged.

- reportShortcutUsed:
Unchanged.

- applyRestore:
Unchanged.

- disableShortcuts/enableShortcuts
Unchanged.

- requestPinShortcuts:
A shortcut cannot be pinned by launcher if it's hidden from launcher. An
exception would be thrown upon requesting to pin a shortcut that is
hidden from launcher.

Bug: 202335257
Test: manual enable feature flag for appsearch integration,
then run atest ShortcutManagerTest1

Change-Id: Ia0e5d31549c9d83efac9bc2a7ea894df425fd5cd