Let prefetch jobs run for apps with active widgets.
We currently can't tell if a widget is displaying data that is
updated periodically. If a prefetch job is scheduled for such a widget,
it should get to run since it'll be helpful for the user. Given what we
know right now, we will mark the prefetch bit as satisfied for apps with
active widgets. We won't relax the force-batching requirement unless we
also determine that the app will be launched soon.

Bug: 194532703
Test: atest FrameworksMockingServicesTests:PrefetchControllerTest
Change-Id: I6d1c17c386a27e70525f72824710739b78e9ad5e