Create a convenience method to get the CarrierService package.
Currently, this just redirects to the existing Intent resolution APIs to
find the first carrier privileged package that declares a CarrierService
implementation for the specified SIM.

Follow-up work will make this a proper API that reaches all the way down
to CarrierPrivilegesTracker, which is slated to become the source of
truth for all carrier privilege related checks. Doing so will also
provide a free caching layer for this API, which should result in
appreciable performance increases.

Bug: 205736323
Test: make
CTS-Coverage-Bug: 205995169
Change-Id: I949de81f2c5b284b14c56b5d8d3ed0b60cb97847