Don't require T to extend Parcelable or Serializable in Parcel
This is because we want to support the use-case where the caller
enforces a specific class type that doesn't implement Parcelable
or Serializable but the child classes written on the payload implement them.

One such use-case that needs this before migrating is
https://cs.android.com/android/platform/superproject/+/master:frameworks/base/core/java/com/android/internal/infra/AndroidFuture.java;l=596;drc=35b2a4bc9e1ba02eb8c43f9574deb63ab797a3a4:
we need to enforce Throwable class but it doesn't implement Parcelable itself,
while its children written on the wire are expected to.

Test: atest -d android.os.cts.ParcelTest
Bug: 195622897
Change-Id: I150416d0cfb0b87ddbaff1041d8d60aa205f6f39