Implementing TIP1.
Tare Improvement Proposal #1 (TIP1) separates supply from allocation.
By separating the two, apps are able to accrue credits regardless of the
balances of other apps, and the limited nature of the resource is
shifted to when apps try to perform actions (and thus consume
resources). We remove the "maximum circulation" that limited the total
number of credits that could be allocated across all apps at one time
and replace it with a "consumption limit" that limits the total number
of credits that can be consumed across all apps within a full discharge
cycle. The consumption limit continues to scale down with the battery
level. This change should help reduce concerns over misallocation.

Bug: 158300259
Test: Spot check dumpsys of tare, jobscheduler, and alarmmanager
Test: atest frameworks/base/services/tests/mockingservicestests/src/com/android/server/tare
Test: atest frameworks/base/services/tests/servicestests/src/com/android/server/tare
Merged-In: I104eef812889ffb543132ae46c7ac869b12c9882
Change-Id: If8b91a56d5bc9a75f21a07b57a55c3b2c86bad12