Remove two timeouts in ResolverActivityTest.
These don't seem to be necessary for the respective tests
to pass, and they were hampering my ability to make changes
to the chooser refresh delay (since they reference the same
symbol). If we do end up needing to reinstate any delay
like this, we still shouldn't have to take a dependency on
Chooser from Resolver.

Bug: 224816815
Test: atest --iterations 30 ResolverActivityTest
Change-Id: Icf806e2babca048726fe33843df4f9f62e4bd138