Further defer loading widgets until it is absolutely needed.
Current implementation parses the provider info from all manifests in
every app the first time getInstalledProvidersForProfile is called,
which causes unnecessary delay on the call-site if only a single package
was requested.

This CL ensures widgets loading are deferred further down the line,
until it is absolutely needed to be parsed.

Bug: 202356231
Test: manual
Change-Id: I793bd818781ae956cd5bc590a315c669d01e7ca6