Transfer remote animation option from an intercepted activity.
While an activity starts, ActivityStartIntercepter may intercept it and
launch another activity. If this happens, we don't have chance to apply
pending remote animation of the intercepted activity.

The pending remote animation originally came from ActivityOption, so we
extract the remote animation from an ActivityOption of the intercepted
activity, and pipe it through to the activity being launched by the
ActivityStartIntercepter.

This also reverts commit 92a727826fd3497c7003daa15262b2b1009656db.

Bug: 222757698
Test: manual. Start YouTube app from the launcher and check animation.
Change-Id: Ice62f4ee3453a06362e3acf820ef22235656845c