Changes to reduce flakiness of TrustTests
Use AndroidTest.xml to disable the double-click on power for turning on
Camera.
While trying to dismiss keyguard, use the back button instead of trying to power cycle the phone in order
to dismiss bouncer. As part of this, also wake the device if it's asleep
since you cannot dismiss keyguard if the display is off.

Bug: 226652325
Test: atest --iterations TrustTests
Change-Id: I7eb920f220dbc8d28d7c2035252365277d51421a