DisplayContent: Don't force-update parent if layering didn't change
A recent change to provide IME support for SurfaceControlViewHost
modified logic for setImeLayeringTargetInner. setImeLayeringTarget
inner also updates the control target, which may depend on the
input target. In the case of SurfaceControlViewHost the input
target can change with the layering target not changing,
and so we need to update the control target, which was done
by extending the conditions on the return at the beginning
of setImeLayeringTargetInner to also include the check for
mLastImeInputTarget == mImeInputTarget. However in standard
app cases, this can produce a scenario where we consider the
IME parent as changed, even when it wasn't (when neither
the layering or the control target update). This CL partially
restores the previous logic by not forcing the IME parent update
when the layering target didn't really change.

Bug: 226338567
Test: Existing tests pass
Change-Id: I8ec9583feca8a0be9c90b0a4a2a382bf915ae4be