Support multi-window handwriting without focus requirement on down
Previously, the handwriting window was only initialized over the focused
window, and re-initialized only when the IME was bound. This meant that
in multi-window modes where the input method client changes without the
IME being re-bound, handwriting would only work on one of the apps.

Here, we initialize the handwriting window over the entire display so
that handwriting can be started over any app. To ensure that handwriting
only starts if the stylus stroke starts within the bounds of the app, we
do a hit test with the start of the stroke before starting handwriting.

Bug: 230381729
Test: manual: test handwriting in split screen
Change-Id: I9296d35e29387ea54afbdeabe2baa26bf6ccbc71