Fix blockable system notifications
Notifications can be non-blockable if their app has the
permission fixed or if the app holds certain roles.

Calculate that on device boot and role change, and use the same
source of truth for all UIs. As a bonus, we can remove some
systemui binder calls.

Test: SystemUiTests & FrameworkUiServicesTests
TesT: verify that 'usb debugging connected' notifs are non blockable
but magnification ones are, in systemui and settings
Fixes: 231662091

Change-Id: I980718ab61196901d976f9ea8ee035eafc021fc8