Return the activity type of any child activity in the Task
App crashed because the activity type was set to UNDEFINED
while the activity was already set to STANDRAD, which is
similar to b/228592780. It happened while the task only
contains finishing activities.

Following commit 33af88e, the activity type should be the
same with the other activities in the Task even the activity
in the Task is currently finishing.

Bug: 233550853
Test: Drag settings from taskbar into split

Change-Id: I1c1f7b67f019bc1c3a84d38bc0ac6e7c188d1e49