PeopleSpaceWidgetManager: log slightly more
We're seeing a weird probelm where Conversation widgets are still
showing content even though the relevant notifications have been
dismissed.

I'm currently assuming that the widget is failing to update for some
reason when the notification is dismissed.

This change logs all the warnings and errors in PeopleSpaceWidgetManager
unconditionally, and includes the widget ID in all of them. Hopefully
this will give us a clue what's going wrong when the notification is
dismissed.

Bug: 230268718
Test: manual; errors logged trying to add widget with no convo notifs
Change-Id: Ie7fdf91dccbb4b02ba4b9ffa9e20ec758abb96b9