Remove Keep Clear Area derived from IME-height on TV
Not all TV keyboards span the entire screen width, and the TV PiP window
should be allowed to be placed in the corner of the screen next to the
keyboard if there is space.
Previously, to make sure the PiP does not cover the keyboard, we used
the IME-height to create a keep clear area along the bottom of
the screen.
But we are changing how keep clear areas are derived from IME windows,
and the IME-height keep clear area is no longer needed.

Bug: 227596282
Test: atest PinnedStackTests
Change-Id: I8a04ba618998d025ddd2dae27bbf9f6ed6a1a599