Fix no announcment when the PIP menu shows on the screen
In the legacy implemenation, the PIP menu is a activity,
so the property of windowInfo, inPictureInPicture is true.
Screen-reader will read the the tile for the new pitcure-in-picture window.

We set this property of the PIP menu  to false because it is not a activity anymore,
however, it ends up that there is no announcment anymore.

To fix it with the simplest way, we set this property to true.

Bug: 236216130
Test: manual test, enable Talback and show the pip menu to see if there
is any announcement

Change-Id: I73e36242db074c747769b38b328a9cdc742fbd90