Fix exception when opening App info on work profile
The root cause is that the target intent was intercepted to
ConfirmDeviceCredentialIntent and ActivityStarter tried to embed
the credential intent unexpectedly.
This CL fixes the issue by, instead throwing exception,
start Activity to the new Task if the parent Task of target
TaskFragment is different than the target Task.

This CL also removes the remaining embedded Task as tracked in
b/234351413.

TODO:
1. With this CL, the app detail info page after intercepted shows
   in fullscreen. It will be embed back to the right TaskFragment
   as follow-up CL.
2. Fix the layout issue of ConfirmDeviceCredential Activity.
   The layout is flaky at the first time to launch device.

Bug: 233578672
Bug: 231972879
fixes: 234351413

Test: atest ActivityStarterTests
Test: atest CtsWindowManagerJetpackTestCases
Test: atest TaskFragmentOrganizerControllerTest
Test: atest TaskFragmentOrganizerPolicyTest TaskFragmentOrganizerTest
Test: atest SplitActivityLifecycleTest

Change-Id: Ibcb40521a5a8374e485932f006049ba676213da1
Merged-In: Ibcb40521a5a8374e485932f006049ba676213da1