Fix constraints for collapsed layout
The missing constraints for the time views would cause the play/pause
button to be laid out incorrectly in RTL mode. Added values to match the
prev/next buttons (same as the expanded layout). Also removed code that
was changing visibility of these in code since the time views will
always be GONE in the collapsed layout

Fixes: 237730543
Test: manual (video in bug)
Change-Id: I6a9e29de550c4bba05d6fc38e3be3798f0b525b6