Inline IMMS#startInputOrWindowGainedFocusInternal
This is a follow up CL to my previous CL [1], which aimed to be a
logical revert of changes mentioned in the commit message.

One thing that CL forgot to handle was

  InputMethodManagerService#startInputOrWindowGainedFocusInternal,

which was mentioned in one of our previous CLs [2].

This CL inlines that method again for simplicity.

This is a mechanical refactoring.  There should be no behavior change.

 [1]: If16ac0de536d9089eb04f6e07b1ee47378124658
      d98017b3ef49cb0d5020548a7fd942c37c0178a8
 [2]: Ic584203c1221fbae17f5e2d8f09e3992df061646
      698113215aafbee50d410003e5e22a9d2186b682

Bug: 234882948
Test: presubmit
Change-Id: I7c775245882815483f29acfd83f53b08a40d07ed