Deletes LifecycleActivity.
LifecycleActivity seems to only be adding a Lifecycle to the framework
Activity class. The Jetpack ComponentActivity already does that and is
actively getting fixes whereas the Activity class is not.

This CL deletes LifecycleActivity and migrates existing subclasses to
subclass ComponentActivity from Jetpack, instead.

Bug: 243558484
Test: Opened and played with the home controls screens and the user
switcher screen. Did not test Wallet, but I have high
confidence that this change is safe given its nature and the manual test
for home controls.

Change-Id: I10f83c71d02e7218a0196b272006dbe6d17156dc