Fix finishTransaction window crop with requestedOverrideBounds
Before, when the requestOverrideBounds is not at (0, 0) when its parent
is also not at (0, 0), the window crop will be different from the actual
visible area.

For example, when enter ActivityEmbedding split with system split.

The requestedOverrideBounds is in screen coordinate, and the relative
position offset is in parent coordinate, but the window crop should be
in the current surface coordinate. Update it to always crop from (0, 0)

Bug: 207070762
Test: manually verify with enter system split with AE split.
Change-Id: I45dad68c93921ae0c7b5f0b46553014780a0c9b8