Delete PanelViewController
Second attempt at ag/19668417. This is not a revert of the revert.
The CL was regenerated from scratch to reduce the risk of merge
similar to the one that caused the regression in the first attempt.
The CL was originally created a month ago, but it was shelved at
patchset 5 to avoid regression risk. To update the CL with a month
of changes, it was simply regenerated from scratch instead of
manually merging those changes into a complicated CL. Patchset 6
was then compared to 5 to identify all of the style-only fixes
that were in the original CL so those could be preserved.

Bug: 242211739
Test: ran atest, manually tested shade, manually verified b/244276156 not present and that HUNs work
Change-Id: I561ce6344ea77d5c55ed83762c543ec1078e64bb