Optimize IMMS#setAdditionalInputMethodSubtypes() a bit for bg users
This is a follow up CL to my previous CL, which enabled background
users to call

  InputMethodManager#setAdditionalInputMethodSubtypes().

One this this CL tries to optimize is the fact that its implementation
is currently invoking

  AdditionalSubtypeUtils.load()

twice when called for background usres.  The double-call is happening
in

  InputMethodManagerService#setAdditionalInputMethodSubtypes()

itseif and inside

  InputMethodManagerService#queryMethodMapForUser() [2].

With this CL, the load operation should be happen only once.

There should be no developer-observable behavior change.

 [1]: Iac778d95846843504f6e00b3b939ef0794013b02
      c82a912f24c0337edab4a4141a0aedc65777227a
 [2]: I4187468076705ac597d680f2f5dc32d7b166da1f
      057ea26ce11ff6bd0b9285b7f4d089e1b08d873a

Bug: 234882948
Test: presubmit
Change-Id: I6eb4201ebd69ca8821cbe74fc69356f68f9bcb5e