Use String[] for IMMS#mNonPreemptibleInputMethods
This is a safe optimization for our previous CL [1], which introduced

  InputMethodManagerService#mNonPreemptibleInputMethods

as HashSet.

Since it is almost always empty for phones and is expected to have
only a few entries for certain form factors such as TV, using HashSet
is most likely overkill.

Let's just use String[] and ArrayUtils.contains() for simplicity.

There should be no observable behavior change in this CL.

 [1]: Id0aaa496ee46532d0e97c236df7e073947ababcd
      2deb6ac35e4a09856e451d9ea0ad21d895be3de8

Bug: 216618758
Test: presubmit
Change-Id: I7269306f56df4d083de31a2d9d5b49856f9a9f5d