Remove InputMethodManagerDelegate#getLockObject()
This is a follow up CL to our previous CL [1], which introduced

  InputMethodManagerDelegate#getLockObject()

to clarify that "mServedView" and "mNextServedView" need to be guarded
with

  InputMethodManager#mH

as a short term workaround.

With this CL, above two fields are moved back to InputMethodManager
like they used to be before we introduce ImeFocusController [2].

This is a mostly mechanical refactoring.  Hopefully there should be no
observable behavior change.

 [1]: I9c072b829d1db1e68b65e766d764ee71cb16e6a2
      a9cba8d709ff0edd508ab9d9139d16d7be9a1376
 [2]: Ib455704fe1e9d243f93190a84f230210dbceac2a
      efbe251506c65ba1a5affe932cb22dad91edb449

Bug: 244504062
Test: presubmit
Change-Id: Ie3446e1d0f62c489195e31fc7f073020884d6cae