Fix bug where switching to guest didn't work.
The cause was that we were using UserRecord.resolveId() which would
return UserHandle.USER_NULL when the record was for a guest. We now use
the raw ID in the UserRecord in this case.

Also, the logic is being moved from UserSwitcherControllerImpl which is
a temporary class meant to help us migrate from the old implementation
to the new and is thus not unit tested to UserInteractor, where we can
comfortably add unit tests in this CL.

Fix: 251740146, 251376342
Test: Issue no longer reproduces after this CL
Change-Id: Ib0cbb23d0f6111af7a2eb98278796c34a3e69209