Let IMMS call isKeyguardSecure() for the correct user
InputMethodManagerService has been using

  KeyguardManager#isKeyguardSecure()

to fine tune when the IME swicher icon should be shown on the lock
screen [1].  However a recent CL made that API @UserHandleAware [2],
which was not reflected in InputMethodManagerService yet.

With this CL, InputMethodManagerService can query isKeyguardSecure()
for the correct user.  A bonus is that this CL removes the direct
dependency on KeyguardManager from InputMethodManagerService.

 [1]: I2ae7fc7676e5b64d8910dbc6833d3042e8d2329a
      c2763dce602218db7997c8279b484d3dbb027e77
 [2]: Ieb7217f5e1dddb65f8987242e675c3dff0a80aab
      7dbca17ffacb53f3b0f27310bb24f6d16b205561

Fix: 251459998
Test: presubmit
Change-Id: Ia29e303fe07fd876ea2ecfb937c1fb5b867a0e89