Remove unused flag DEXOPT_ONLY_SHARED_DEX.
It's not set anywhere except in tests. History shows it was implemented
in 1aa1c006 and partially reverted in cc3e245e. However a flag
refactoring in 988445fc got in between, so some parts of the flag
handling code wasn't cleaned up in the revert.

The flag is removed keep parity with ART Services, where it is not
(currently) planned to be supported.

Test: atest DexoptOptionsTests
Bug: 251903639
Change-Id: I09b3eb09636c2add529e1c77e938e073d634fe50