Not fetching MediaProvider from Clone Profile
Ideally clone profile should not have MediaProvider instance, even if
have the MediaProvider avoid fetching it and redirect to owner profile's
MediaProvider. This change is in reaction to bug where MediaProvider is
being installed in clone profile, adding this safe check so that even if
MediaProvider is installed in clone profile it is never used and
redirected to owner's MediaProvider. This check would be present to
ensure fallback from errors.

Test: atest CtsAppCloningHostTest:com.android.cts.appcloning.AppCloningHostTest#testCrossUserMediaAccess
Bug: 236121588
Bug: 254302429
Change-Id: I9d5bcb3dcbcbf1559b45841296566c900738bf1a