1/N sideFPs bouncer, separate altBouncer callback
The AltBouncer callback should only
handle events relevant to any alternate
bouncer (ie: show and hide). Other callbacks
like setQsExpanded and onTouch should be moved
out to a different callback since they're not
relevant to all alternate bouncers.

This CL also converts UdfpsKeyguardViewController to
kotlin and, when the modern bouncer is enabled,
uses the BouncerInteractor for the bouncerExpansion
to determine whether to show the UDPFS affordance
or not.

Test: atest UdfpsKeyguardViewControllerTest
Test: atest UdfpsKeyguardViewControllerWithCoroutinesTest
Test: wallet activity shows udfps affordance
Test: tapping home controls affordances
shows udfps bouncer
Test: tapping notification w/ intent shows
udfps bouncer
Bug: 237362467

Change-Id: Ibb9b55e6b1a629e28f523ac62058f7e5cf40faed