Revert "libandroidfw: use map for more efficient indexOfString()"
This reverts commit e97e96914bbaabb78906b8d6dc05c3eb22b4e0ad.

Reason for revert: b/253420843

This CL causes memory regression as described in above bug. To speed
up the lookup from resource string to resource ID, the original CL
introduced std::unordered_map for the purpose, unfortunately, that
adds much memory footprint due to the number of key strings per
process. We will need to revisit the solution with a more memory
efficient approach.

Change-Id: I9c6754d361f091e963d4bebea0e698e72e221e51