Rename IInputMethodManagerGlobal to InputMethodManagerGlobal
This is a mechanical renaming for classes outside
android.view.inputmethod package.

For non IMF classes, let's expose InputMethodManagerGlobal rather than
IInputMethodManagerGlobal.

In the next CL [1], we want to merge IInputMethodManagerGlobal and
IInputMethodManagerInvoker into a single class, and this CL is just a
preparation.

There must be no observable behavior changes.

 [1]: I310a33f27b876419b6a3058202dcbd663b541079

Bug: 234882948
Test: presubmit
Change-Id: I9b728560369179d5ee17c3570a2543b49b3f14ce