Clarify required permission in IMM#getInputMethodListAsUser()
This is a follow up CL to our previous CL [1], which exposed

  InputMethodManager#getInputMethodListAsUser(int userId) [2]

as an @TestApi.

One thing to clarify is that INTERACT_ACROSS_USERS_FULL permission is
required when and only when "userId" is different from the one of the
calling process.  To make it clear, this CL adds "conditional = true"
like other @RequiresPermission usages in InputMethodManager.

This is just about (Test) API annotation.  There must be no observable
behavior change.

 [1]: I4b73d9374d7a2ce4c9dfc48682a43d5d48b39042
      c6838e2713360cc3fd6aec2108565a17534d1fc3
 [2]: I5b50b5fe143c74c87b331bda3e5bcc4d6248436e
      0787d3f85e2e94ca7fe29f1c5599408e5c50fd33

Bug: 256239860
Test: presubmit
Change-Id: I14867e528d4292f1f8713b1d9b491fc5ed6c1abb