Remove an unused param from IMMS#showInputMethodPickerFromClient()
This is the 2nd attempt to submit this cleanup [1], after a revert [2] due to some build failure.

This is a follow up CL to our previous CL [3], which introduced

  InputMethodManagerService#showInputMethodPickerFromClient(
          @NonNull IInputMethodClient client,
          int auxiliarySubtypeMode, int displayId);

Somehow the "client" has never been used, which is fine because the
actual caller verification is done by requiring WRITE_SECURE_SETTINGS
permission.

Let's remove such an unused parameter for simplicity.

There must be no observable behavior change.

 [1]: I2f27160cbe7d83201c8f52dbb6dcb774a5b8ae43
      ae086b3924309fcb1b156ae26f843dd40471116c
 [2]: I343828ad83d50436211004753f1b9ba1b13b60ca
      41389657ee9415ea2369be70c77e6558ce374cf9
 [3]: Ic7d7c5a7ad8005a3fbd9d1c1b73e3c5a39a07001
      bf36ccc7de8dcec478910ffbbd2dd7c313157824

Bug: 234882948
Bug: 256952514
Test: presubmit
Test: lunch aosp_bramble-userdebug && make -j
Change-Id: Ia536d8793f0022179dfafd1cbf21091537375992