Allow awaitLux to fetch a value after destroy.
The logs are showing -1 lux values at a higher rate than expected (~10%). This may be because the lux probe is being destroyed before a value is ever reqeusted since the scheduler and auth callbacks are confined to the same thread but use different handlers. This allows destroy to be postponed until at least one value has been recorded.

Bug: 253318983
Test: atest ALSProbeTest
Change-Id: Id304bf502bcea9dc313f21253e117adb96324e10