Revert "Add permissions team to sensorprivacy.proto OWNERS"
This reverts commit 4dfa4273f68d7a78c2fb2c713f187be6688e867e.

Reason for revert: 

The CL causes merge conflicts for the updates on AOSP. 

I first tried to cherry pick to aosp. But as "ewol@google.com" is no longer a valid email address, the cherry pick failed.

So I revert this CL and move the correct content to it's AOSP counterpart.
https://android-review.git.corp.google.com/c/platform/frameworks/base/+/2346822

Change-Id: I68f98b963e3ea97d71beb5504d8682ac093edd0f