[Bouncer] Do not send message if face auth...
Succeeded.

When we authenticate the face, the KeyguardIndicationController checks
to see if the bouncer is showing and will set message to "Swipe up and
try again." if so. Ensure that we only show "Swipe up and try again." when face auth
fails.

Also ensure that message is removed when bouncer no longer becomes
visible.

Fixes: 259007288
Test: Enroll in FP and face. Open bouncer and auth via face. Turn off
and on and open bouncer again to view initial message.
Test: Unit Test

Change-Id: I8ee56d27ff60f32e696ac100a64d0553ce3c77e0