Fix delay on TYPE_TASK_FRAGMENT_VANISHED
Before, we call sendTaskFragmentVanished after the removeImmediately,
which may not have any following activity paused/resumed to trigger
relayout to dispatch the pending event to the organizer.

Now, we call mWindowPlacerLocked.requestTraversal() to ensure the
vanished event will be dispatched.

Without this change, the splitInfo callback can be stucked on waiting
for empty TaskFragment vanished event.

Bug: 262666213
Test: atest CtsWindowManagerDeviceTestCases:TaskFragmentOrganizerTest
Test: atest WmTests:TaskFragmentOrganizerControllerTest
Change-Id: Ibf76af205a5fc91400bdd5fc967a0d13591cfe92