CameraDeviceImpl: Follow through reporting device errors even if the remote device was never initialized.
It's possible for setRemoteDevice to be called *after* libcameraservice
calls onDeviceError. It would be risky to implement changes to ensure
setRemoteDevice gets called first , so use a simple boolean flag to track
whether the remote device was initialized. If it wasn't, then we can't
assume the user isn't interested in device errors - that is only true if
the device was closed.

Bug: 263429133
Test: Ran CtsCameraTestCases on cuttlefish
Test: StreamedAppBehaviorTest#appsInVirtualDevice_shouldNotHaveAccessToCamera
Change-Id: Id1990dcb17e5ec2afc562bf76e410b58a49dcf77