Revert fs-verity for all APKs
Due to the system health concern of cold app start time, revert the
change that applies fs-verity to all APKs.

There's an ongoing work to improve performance (b/233247259 for
dm-verity, but the idea is applicable to fs-verity, if works) and may
reduce the impact to an acceptable level. Once that lands, we can try to
re-enable again.

Bug: 249158715
Test: atest ChecksumsTest ApkVerityTest
      CtsAppSecurityHostTestCases:android.appsecurity.cts.ApkVerityInstallTest
Change-Id: I1ec7273bbb255745a1854ceb7af15bfa2738e265