Implement the runtime APIs
Implement the following APIs:
- finishActivityStacks
Finish a set of ActivityStacks.
It is useful to expand the primary container.

- invalidateTopVisibleSplitAttributes:
Update the container actively and trigger the SplitAttrbutes
calculatator function.
It is useful to update SplitAttributes regardless of
device and window state change.

- updateSplitAttributes
Update the SplitAttributes of a specific split and also
overrides its default SplitAttributes.
It is useful if an app doesn't set its SplitAttributes
calculator function, but want to customize SplitAttributes of
a split pair without updating the rule.

Bug: 263565444
Test: presubmit

Change-Id: Icc502f5e106d380ed1e5656445fbd20a1a391c8e