Fix incorrect behavior of back button in kids mode
Previously, we created a root task with its size excluded the nav bar
for kids mode and put all tasks in this root task execpt for HOME task
since the kids mode team want the kids launcher to be fullscreen.

The issue can be reproduced as following steps:

1. Task Hierarchy when Start App1
  - Kid's Root Task
     - App1 Task
  - Home

2. Task Hierarchy when Start home
  - Home
  - Kid's Root Task
     - App1 Task

3. Task Hierarchy when Start App2
  - Kid's Root Task
     - App2 Task
     - App1 Task
  - Home

4. Task Hierarchy when user press BACK key. The App1 task will be
   brought to front instead of Home
  - Kid's Root Task
    - App1 Task
    - App2 Task
  - Home

Solution:
We now also put HOME task in the root task but set it's bounds to
fullscreen size and set the window crop of the root task to fullscreen
size so that HOME task can extend to fullscreen size.

Bug: 266507175
Test: 1. launch app1 and press Home
      2. launch app2 and press Back
      3. Home should be brought to front
Change-Id: Ic6d5f65f277a4b73062d51fc67dcc300ab1ce15b