Fix App List pages flickers
When there is a package change event happens on the device, the page is
auto refreshed, which caused the flickers.

Before this change, the app list item will be re-create for every app
list refresh, because System.identityHashCode(this) is part of item key.

System.identityHashCode(this) is added to fix an issue in the
AppListSwitchItem.

To fix this issue, removed remember in the AppListSwitchItem to make
sure TwoTargetSwitchPreference always has the latest values, so
System.identityHashCode(this) no loner need to be part of item key.

By removing System.identityHashCode(this) from the item key, the flicker
issue is fixed.

Fix: 267699436
Test: Manually with Settings
Change-Id: Iabf0c56271c5486272c921a4b4020f9e3d2b984b