Stabilize overview/quick-switch transitions
When done in rapid succession or at some bad timings,
it was possible to get stuck either with touch being
ignored or with some windows incorrectly invisible.

This is fixed primarily via tracking of task state in
the recents transition animator to account for many
changes being merged into one animation. This includes
a task-switch being interrupted by a re-entering gesture
as well as re-opening of the live-tile due to unsolicited
intents (eg. notification permission activity).

Without detailed tracking, the WMCore state and the
recents animator state would fall out of sync.

Second, make sure that visibility-changes resulting from
`setFocusedTask` calls are done in a transition. This
API was just missed before so Shell wouldn't be notified
of some visibility changes.

Bug: 269691848
Bug: 269691226
Test: existing tests pass. Also manually do rapid quick-switches.
Change-Id: I0b465b11ea316fc670a1ba121e61535e22108be5