Indicate potential incomplete transition
To catch the case such as when an activity is still pausing, and
a caller sets transition to ready too early. And then when resuming
next activity in the same task, its visibility will be updated
without an associated transition.

Previously, the visibility will be skipped because the activity
is a child of the task in transition participant but itself is
not in participants.

Bug: 269732484
Test: TransitionTests
Change-Id: I9a184b97974f337fb123e11853bea50ace9bf843