Add logging to help debug PipPinchInTest failures
PipPinchInTest passes both locally and on forrest
on top of the last known good build. But there seem
to be occasional failures happing according to some
of the earlier runs, and they can be hard to reproduce.
Thus, we are adding extra logging in case there are any future
failures at the postsubmit stage.

Bug: 269505548

Test: atest WMShellFlickertests:PipPinchInTest
Change-Id: I302b498dfb849084cd581652ced85a648fb49273