Add a TaskState tracking mechanism to Recents transition
This is a further re-arrangement. No logic should change.

This intrudeces a TaskState class which can track some
properties of a task which recents animation cares about.
This change is separated from the one that actually changes
logic to minimize mixing of rearrangement and logic change.

In order to have the right information for the task-states
(in particular, both the task surface and the wrapper leash),
this had to localize the RemoteAnimationTarget wrapping into
the recents wrapper. This should do the same thing though as
its using the same filters.

Bug: 269691848
Bug: 269691226
Test: no logic changes... existing tests pass
Change-Id: Ibc208bd147d84069ae30b010768a6112c57c09b1