Stop blocking broadcast receiver thread
Stop blocking broadcast receiver thread during network operations.

This reverts part of a behavior change from commit 21c21e53e6b.
That commit switched a BroadcastReceiver over to using a specific
handler (X) for some long-running work. That isn't the same behavior as
posting the work to handler X and returning immediately, and it can
cause an ANR if the work takes too long.

The behavior has been changed back to explicitly post work to the
handler and return.

Bug: 271951363
Test: build / boot / treehugger
Change-Id: I9aeaf200d19de97b5be909678aab8d5421cd11a4