Adding experiment flags for optimization part of autofill for all apps
For experiemnt, we added two flags:
1. Enable heuristic check on important views. This flag is default to
   false. If this flag is enabled, the imeAction check will also be
   applied on important for autofill views. We want to see if enabling
   this would decrease non-autofillable requests received by autofill
   providers, also check if there is any regressions.

2. Flag to indicate whether to have view type check on not important
   views. It's default to enabled. We wnat to see if disabling this
   would cause any regression.

For potential bug line, we will be only checking package partially
denied if package is not fully denied. In current logic, if package is
fully denied, it will always be partially denied. Because whether it's
fully denied is checked by if "package:;" is in denylist and partial
denied is checked by if "package:" is in denylist.

bug:265186343

Change-Id: I613d3a76af1252903cd5a5b12746e1ae8ad33615