Fix lockscreen shortcuts not showing on split shade
The issue was that qsVisible was wrongly being set to `true` when on
split shade, even when it was not visible.
Lockscreen shortcuts only show when QS are not visible.

To check whether QS (or QQS) was visible, we were checking:
1 - qsExpansionFraction > 0
  1.1 Split shade: true when full QS showing
  1.2 Single shade: true when full QS showing
2 - qsPanelBottomY > 0
  2.1 Split Shade: ALWAYS true. Even if QS not showing
  2.2 Single shade: true when on QQS or QS

Since on split shade we can always rely on qsExpansionFraction, because
there is no QQS, we changed the condition to ignore qsPanelBottomY
when on split shade.

Fixes: 272448562
Test: Manually on-device:
      1 - Enabled teamfood flag
      2 - Homescreen -> Wallpaper & Style
      3 - Lockscreen -> Pick 2 non-none shortcuts
      4 - Lock the screen
      5 - Verify shortcuts are visible both in single and split shade
Test: QuickSettingsControllerTest.java
Change-Id: I9adc0c4625aadcc729b484c453c6ef7d17013fc9