Grant MANAGE_USERS access to Traceur
This change updates the privapp allowlist to grant the MANAGE_USERS
permission to Traceur. This permission is needed to query admin user
status, as Traceur shouldn't be able to start if the current user is not
an admin.

Test: Using ABTD, apply this change with ag/22119816 to verify that
      Traceur still works as intended (opening app, tracing, etc.).
Bug: 262243665
Bug: 262244249
Change-Id: I8e2174065b686c052cb080b3590ea4d89e7a7783
Merged-In: I481c95fed4a364239f0671ed1b4d9bfaa0b07bd6