Use testable components to cleanup display tests a bit.
Second attempt. TestableSettingsProvider currently doesn't understand
the special UserHandle values, so we need to use the same user value in
the test as we do in the code under test. If we don't, it will read
through to the real settings value, making the tests unnecessarily
depend on external state.

Fixes: 274430330
Test: atest DisplayPowerControllerTest DisplayPowerController2Test
Change-Id: Id4568924aa67a3be6e0c0169a168b9241e1e510c