Make registerDexModule a nop regardless ART Service is enabled or not.
ART Service doesn't support that method, but the callback should still
be called, for compatibility.

This also disables the method in the legacy code, to align behaviour
when ART Service is enabled or not.

The main problem with this API is that it doesn't take a classloader
context, so it is hard to produce dexopt artifacts that can be
succesfully loaded by the runtime later. This problem exists in the
legacy code as well.

Test: atest DexModuleRegistrationTest
  with and without dalvik.vm.useartservice=true
Bug: 274096060
Bug: 37290820
Bug: 251903639
Change-Id: Ie1510268a721e846408b40819cd681063cfbb485