Do not separate flow into two.
In the logs of the bug report, bouncershow(true) and bouncershow(hide)
is called 50 ms apart. The theory is that the ordering of the flow
collection is not deterministic because bouncershow is separated into
two separate flows with a map.

Also added a log in the bouncer repository when the bouncer is
showing/hiding.

Fixes: 274608304
Test: Open and close bouncer from occluded state.
Change-Id: I57595b76af9385bb83ab007f25347a5566e9903a