Do not override finish activity behavior on back event
... if the activity was not started from Launcher.

The launcher activity was started from notification and was not
removed while swiping back. Another activity instance was added
on top of the same task when the app was started from Launcher
because the intent that started the task and the intent that
started from Launcher are different.

Also adding unit test to verify task is moved to back if the root
activity was started from Launcher.

Bug: 272723475
Test: atest ActivityRecordTests
Change-Id: I97ccce3a7491669ab6b649b6e1f053beae3828f0