Properly sets the display id on Autofill UIs.
By default autofill is shown on display 0, but in order to support
visible background users, it must show it to the proper display.
Previously, it was using the display assigned to the user, which
would work in the majority of the cases, while this CL explicitly
uses the display associated with the activity (that would support
autofilling in a virtual display, like a panoramic display on front
passenger seat).

Test: atest --user-type secondary_user_on_secondary_display CtsAutoFillServiceTestCases:android.autofillservice.cts.dropdown.LoginActivityTest#testAutoFillOneDatasetAndSave
Test: atest CtsAutoFillServiceTestCases # on phone
Bug: 271031908

Change-Id: I346110063aa346e0b7b471dd8e0ca37330fff5e7