ViewModel+Binder for NotificationStackScrollLayout
This change lays a foundation for more notifications code to be migrated
to the Recommended Architecture. Right now, it primarily serves as a
means to integrate the NotificationShelfViewBinder into the codebase
without needed the NotificationShelfController interface any longer. As
more and more child Views of the NSSL are migrated, the NSSLViewBinder
will continue to grow.

If we ever decide to migrate the NSSL itself, then code from the
NSSLController will be moved into the NSSLViewBinder under a process
similar to that implemented for the NotificationShelf.

Bug: 271161129
Test: atest SystemUITests
Change-Id: I8f605f8d7b4722d3c2e065c7b37cc4f23e6cc99a