Update RoundableState logs
The dump of the NotificationShadeWindowControllerImpl contains its view hierarchy using the `toString()` method of the `View` class. We want to keep it, because it contains some basic description of the View state, visibility, clickability and the view's position inside its parent.

The dump of the NotificationStackScrollLayout knows more about our custom views and their custom properties.

This CL updates the RoundableState logs to follow this pattern.

Fixes: 278509751
Test: dumpsysui NotificationStackScrollLayout NotificationShadeWindowControllerImpl
Change-Id: Id64fc6e8ed692c530aace95eb3036c0b37af647a