Fix connection enumeration not accounting for discontinuous IDs
The data structure that contains mConnections is a sparse array. This means that you are allowed to insert devices with unique input IDs that are not continuous. However, when trying to perform the search for a specific input ID, the current logic assumes that the IDs are numerically continuous, causing issues (see ag/22753602).

Bug: 278628657
Test: atest TvInputService
Change-Id: I0311ac994feca73e4216744a49f1731c8e353ca1