Prefer the VCN subscription when filtering for CBRS
This logic only matters in cases where there are exactly 2 subscriptions
with the same `groupUuid` and at least one of them is opportunistic. In
those cases, SystemUI was preferring to track only the subscriptionId
associated with the `activeDataSubscriptionId` (per telephony) and
filter out the other subscription.

With this change, we now track the `vcnSubId` of the default network.
`ConnectivityRepository` exposes the subscription ID of the current
cellular VCN if it exists. In the case where one of the opportunistic
networks is a VCN, we will prefer _that_ network rather than the current
active data subscriptionId.

The case that whis becomes relevant is any time the active data
subscriptionId differs from the subscription id reported by the
VcnTransportInfo. This case should only happen when telephony tries to
switch the active data subscription and finds it not validatable. What
happens then is that the VCN is still connected and we will continue to
show the subscription for that connection rather than the active, which
should switch back.

Test: ConnectivityRepositoryImplTest
Test: MobileIconsInteractorTest
Bug: 222699760
Change-Id: I486ee0c323556407441de36284a1c22d74f3b836