Use getSystemService(String) in PermissionEnforcer.fromContext
Depending on the variant of Mockito linked into a test, it may or may
not support the mocking of final methods. fromContext previously used
Context.getSystemService(Class) which is final. It means that for some
services, the mocking had to target this method; while others were
mocking the underlying methods (getSystemServiceName(Class) and
getSystemService(String)).

Use getSystemService(String) so that the behaviour is identical
regardless of the variant of Mockito used.

Test: atest FrameworksUiServicesTests
Test: atest TeleServiceTests
Test: atest FrameworksServicesTests:com.android.server.lights.LightsServiceTest
Test: atest EnforcePermissionTests
Change-Id: Ic2f8874088c0a8586755c1d3204f5ba39a377096