Add BiometricPrompt implementation with class 3 support.
This is a prerequisite change to restore the delayed fingerprint sensor logic that was removed when strong coex support was postponed last year.

It is not used in the new flagged BIOMETRIC_BP_STRONG SysUI code, but can not easily be flagged and potentially impacts both implementations. This cl is not standalone and requires the corresponding SysUI changes in the topic.

Bug: 272510026
Bug: 251476085
Test: atest AuthSessionTest BiometricServiceTest
Test: manual (see other cls in topic)
Change-Id: I474329496c87bcbedfff43024d8c3bfde7824dd7