Fixing shade not collapsing after rotating from split shade
Issue was happening because when swiping to overscroll, mExpandImmediate in QS was set to true and was never reset. It should be true only when QS are in motion so the solution is to reset it after any motion is finished. Calling setExpandImmediate() twice is safe because if value doesn't change it doesn't do anything.

Bug: 282095041
Test: Open split shade -> scroll down to do overscrolling -> rotate to portrait -> QS should be collapsed
Test: NotificationPanelViewControllerTest#onShadeFlingEnd_mExpandImmediateShouldBeReset
Test: e2e test: SplitNotificationShadeRotateToPortrait
Change-Id: I0663bc28a48b2092546a2cb9a99993320fb7c8e4