Update the java doc to match the behavior
The javadoc of getProvisionedVpnProfileState only describes the
VpnProfileState will be null if if there was no profile
provisioned by the calling package. The implementation requires
the VPN profile being prepared so that the owner UID and package
name are well set up. Design is either changed to refer only the
package name information and check provisioned VPN profile from
VPN profile store, or correct the javadoc. In practice, a VPN app
that provision a VPN profile without starting it should know the
VPN status which is disconnected. The API should be useful after
starting the VPN profile.

Update the doc to match the behavior.

Bug: 280021394
Test: m
Change-Id: I80322ff296661972a442d799812751d5411b665b