Make ScrollFeedbackProvider APIs Public
This change cleans up the existing ScrollFeedbackProvider APIs and
makes them public. As we plan to use these APIs in Compose as well, this
change adds a variant for each API method that takes the raw motion
properties (instead of a MotionEvent, since MotionEvent is not
necessarily accessible from Compose).

Unit tests have been added to the new APIs. These tests are mostly the
exact same tests as the ones existing for the MotionEvent-based APIs,
just with different arguments.

The APIs are launched as a flagged API.

Bug: 299213080
Test: atest HapticScrollFeedbackProviderTest
Test: CTS tests added

Change-Id: I6de611cd34d103ee4f208d421c31b7a3233f2d7c