Update the visibility of split to invisible
DefaultMixedHander handles the going-home request while split screen
in the foreground. mSplitHandler#onRecentsInSplitAnimationCanceled()
is invoked and falls through pipTransition while going home.
The visibility of the split is still in visible state even when the
launcher app is on the top and makes the state of split abnormal.
This CL updates the visibility of split to invisible to prevent one
of split empty. Also update the setReparentLeafTaskIfRelaunch for
relaunch case.

Bug: 285849776
Test: Enable the 3 button
         1. Make a split pair in foreground  ( e.g Map app and gmail)
         2. Click the home button and make Google Map into PIP mode.
         3. Click the expand button to make Map into fullscreen mode.
         4. Drag and drop gmail app to create split with Google Map.
Change-Id: Iabe85b9d0eceb1a093618dde023903e97db1aa28