Do not kill dream process when stopping dream
The DreamController currently force stops the dream with
ActivityTaskManagerService.removeRootTask. Unintuitively, that method
kills the client process if the dream task is the last task in the
process.

We want to keep the dream client process warm and leave it up to the LMK
to decide when to kill it.

In this CL we pass an IAppTask to the DreamController, which allows it
to stop the dream task without killing the process.

Bug: 281960746
Test: atest DreamManagerServiceTests
Test: atest DreamOverlayTest
Test: atest DreamServiceTest

Change-Id: I83f267657812cb776fc65919bc0f5727fdc15b30