Set position during TRANSIT_ENTER_DESKTOP_MODE transition
Currently, relayout runs during the TRANSIT_ENTER_DESKTOP_MODE
transition and sets the surface's position in startT and finishT. This
change overrides that by calculating and setting the surface control's
position to the correct value.

We also introduced the MoveToDesktop animator since we previously had an
animator running at the same time as the enter desktop animation that
was also positioning the task which meant we had to pass in a lot of
information over to the transition handler. Now, we only pass in the
MoveToDesktop animator which holds all the information we need.

Test: Drag to freeform and check that there is no flash
Bug: 277780635
Change-Id: Iaed5191f9d56375c42c32e9afce3ffc54ca5f351