Respect always unlocked displays when keyguard is shown.
When keyguard is shown on the default display, all secondary displays
report mHasContent=false, even if these displays are created with
FLAG_ALWAYS_UNLOCKED (in practice this means that these displays
belong to an always unlocked virtual device).

Setting mHasContent=false on such displays is wrong - any windows
there are visible and interactive. This also does not propagate
display changes to these windows, which causes a bug when the display
is rotated.

Fix: 285141170
Test: m
Change-Id: Ief1fa2db10c234ba179ba741434ce3d6ebef290e