Use the UNKNOWN_ADJ as the cached adj in cycle recomputation
This reverts the behavior introduced in a previous commit 75f3875,
where if a cached app binds to a non-cached app, and this cached app
forms cycles with other cached apps, it might end up bumping
the oom adj score to the non-cached state. So now we'll be using
the UNKNOWN_ADJ. And as we're doing cached oom adj assignment
after that so we won't be running into the unassigned adj problem.

Bug: 286750917
Test: atest MockingOomAdjusterTests
Change-Id: I9284d3d6318dccac1460adae12f24f89bd85b88e