Exclude wake lock and running from tracing.
The wakelock bit races with suspend causing the system property writing
to fail and crash system server. There are better alternatives for this
information in traces via activity manager. The 'running' bit doesn't
providing meaningful information in the way it is in traces so is being
removed to reduce confusion.

Test: atest FrameworksServicesTests:BatteryStatsHistoryTest
Bug: 286017503
Change-Id: Ied906e7ff9e5fe7c141da2c2b85dcef9ca78e027