Skip null surface when validating transition
If SystemUI is died somehow before finishing a transition, the
validators may run after SystemUI is restarted and a new transition
is finished. Then the validators could contains some outdated state
such as the windows were detached (especially wallpaper is died
with SystemUI). So check if the surface exists to avoid crashing
SystemUI again.

Bug: 287539388
Test: Repeat kill SystemUI when running transition animation.
Change-Id: I756fa0bae8b2d64487bc158b933ed21246d3b506