Post reportDrawFinished to the UI thread if running in system
If VRI is running in the system process, the call to finishDrawing will
not result in an IPC call and therefore will run on whatever thread
SurfaceSyncGroup finished on. This is normally going to be the
RenderThread. This could result in a deadlock because the finishDrawing
is now blocking any additional calls to RT, which is unexpected
behavior.

Instead, post to the UI thread when the VRI is running in system process
to ensure RT is unblocked.

Test: Presubmit
Bug: 286355097
Change-Id: Id8bd8eeced99f5228d7655382e0c043d03e0c196