Replace isWeaverAvailable() with getWeaverService()
isWeaverAvailable() has to be called before weaverVerify() and
weaverEnroll(); otherwise mWeaver might still be null.  However, this
was not enforced and was easy to forget to do.

This CL therefore changes the design slightly to make weaverVerify() and
weaverEnroll() take IWeaver as an argument.  To get the IWeaver,
getWeaverService() is called, taking the place of isWeaverAvailable().
Note that this is analogous to LockSettingsService#getGateKeeperService.

Bug: 287705522
Test: Treehugger
Change-Id: I70c26232a9b300a7e6e2c815db6c143836c42ec3