Remove unnecessary animation cancellations when exiting low light
This CL removes the unnecessary cancellation in the transition
coordinator and also scales the dream overlay exit animation from
lasting >1s so that a self-cancellation isn't needed after the view is
no longer visible.

Also greatly simplifies the method we use to truncate the long
animations.

Test: atest LowLightDreamManagerTest LowLightTransitionCoordinatorTest TruncatedInterpolatorTest
Test: manually verified that enabling low light doesn't crash if
tapping during the animation
Bug: 285666217
Fixed: 285666217

Change-Id: Ib0dfe4fe4d030888607e1f03d57f1a084a5fba10