Added additional check on bouncer to check if valid PIN length is stored on device or not
This CL aims at adding an extra check on the bouncer which ensures that
a valid PIN length is stored on device. This provides an extra layer of
security to prevent any accidental lockouts of the user.

Bug: b/284355686
Test: Tested by installing the bouncer on local device and checking if
I am able to unlock the device as expected by setting a 6 digit PIN, and
a non 6 digit PIN as well. Restarted the device as well. Also, tested all CUJ's (b/290641061)

Change-Id: Ie1bb17e6d0b07ed0a0f696a074a94fca8df3991f