Make InputDevice getAssociatedDisplayId part of the test API
While investigating test flakiness, a case where input events
could be dispatched before an input device is associated to a
display has been identified. The issue is that this association
is asynchronous, so we need a way to ensure that the change has
gone through. A way to do this is to monitor the input device
associated display through the API, but at the moment it was
hidden, so this CL will make it available to tests.

Bug: 288321659
Test: No functional changes
Change-Id: I6b9535c160683f7865e766f80fad00dfa4ae52ad