Keep windowing mode by default when removing TDA
This CL lets a task in to-be-removed TDA keep its windowing mode if it’s
reparented to new TDA and  the task is set to FULLSCREEN explicitly.
If it’s reparented to non-TDA launch root (e.g. split root) or the task
isn't set to FULLSCREEN explicitly, we still clear the windowing mode.

For example, let’s say we have two freeform-first displays, launch a
fullscreen task in the display A and remove the display A, which results
in reparenting the task to the display B.
In this case, we want to keep the task in fullscreen even on the new
display B. If we clear the windowing mode, the task shows up in freeform
by inheriting the display B’s windowing mode.

Bug: 290567729
Test: atest MultiDisplayPolicyTests
Test: Launch app in external display, make it fullscreen, and disconnect the display
Change-Id: I124d97948c4b7e0dff47aed2944858d4c03ad384