Catch NPE when reportInstallationStats
The scan result is null for installExistingPackageAsUser(), becuse
it is installing a package that already exists, there is no scanning
or parsing involved. We don't handle this case before and catch
the exception temporary, this change also catch the NPE here. There
is a follow up bug to handle it correctly.

Bug: 298693819
Test: ManagedProfileProvisioningCrossProfileAppsTest#
newlyProvisionedManagedProfile_requiredCrossProfileAppsArePresent

Change-Id: I07d5140785b700a0e9934a0150f709c234dfe0f7