Expand @FlaggedApi(FLAG) constants in API signature files
The auto-generated Flags.FLAG_NAME constants are difficult to review in
API tracking files. metalava will expand annotation arguments if

  (1) the field declaration is known to metalava, and
  (2) the constant is not part of the API surface.

The auto-generated constants are hidden, so not part of any API surface.
This satisfies (1).

This CL adds the auto-generated sources to metalava's input. This
satisfies (2).

Example how how this CL modifies the API signature files:

-    method @FlaggedApi(Flags.FLAG_SCROLL_FEEDBACK_API) public boolean isHapticScrollFeedbackEnabled(int, int, int);
+    method @FlaggedApi("android.view.flags.scroll_feedback_api") public boolean isHapticScrollFeedbackEnabled(int, int, int);

Bug: 297881670
Test: m checkapi
Change-Id: I3259a2f4a114eb8c6eaff0bdbe151f6caef414aa