Add PENDING_AUTH state to AutofillManager
This state allows to keep track from when the user selected the inline suggestion to when we completed the authentication (e.g. biometric prompt or bad password check).

While in this state, we should ignore the changes in view visibility (and potentially other changes related to re-layout) because the user already selected to apply the inline suggestion for that view state.

Specifically this helps with cases when 3P apps do custom visibility change logic when activity goes onPause(). For example, apps can hide EditText for better UX, but it's not a real onPause(), so we should not treat it like one.

Another scenario is a manual password selection via PasswordManager - it takes user to another activity to choose the right password, than provides it to the app - but app has hidden the fields. This fix resolves such case as well.

Verified this change is correctly guarded by the trunk flags by running:
$ adb shell device_config put autofill android.service.autofill.relayout true

Change-Id: Icb0eeecdd037946b1f471e198674fb3be32a7ae7
Fixes: 294330426
Fixes: 295262306
Bug: 238252288