Exclude overlay when reporting top non-finishing container by default
The overlay container is always on the top, and don't split with other
containers. Thus we need the topmost non-overlay container for most
scenario.

This CL also consolidates the usages of #getTopActiveContainer to
TaskContainer#getTopNonfinishingTaskFragmentContainer since it's
the only usage and we usually finish a container after it becomes
empty.

Bug: 243518738
Test: atest OverlayPresentationTest

Change-Id: I1c22a372e3f2c1f31b6ac13c79ddc387b8c253b0