Rename focusGrantToken to inputTransferToken
Hosts can get a token for the embedded window to make focus request call
in system server. The token is just a Binder object that is used to look
up the embedded window in system server. Rename to inputTransferToken
since we can reuse this token for other input transfer requests.

Bug: 303254234
Test: SurfaceControlViewHostTests
Change-Id: Ibf7712a54629e8c5cf9095a08c1cfd501aa4ef78