Run normal finish for entering standalone recents activity
(This is to reland with fixing FallbackRecentsTest#testOverview)
After adding the behavior to finish recents transition when recents
activity is launched on top of home, the case of launching home
again no longer finishes from shell:merge -> recents:onTasksAppeared
-> shell:finish.

The behavior will be like a regular launch, so just make home invisible
after recents activity is launched. And when launching home again,
it will be a new regular task-open transition which is handled by
DefaultTransitionHandler.

Bug: 302703228
Bug: 294029838
Test: WindowInsetsBehaviorTests

Change-Id: Ibcb68f12172379e9e79a9774aa27c5d3d966c6ba