Attach background surface to transition root by default
I02cd3a232350a1dc74e755f9445179ccb216507f changed where the
background color surface is attached to from its transition root
to its task display area to avoid the surface from hiding one of
the split tasks unexpectedly.

However, this doesn't work in freeform mode because when an
activity-to-activity transition happens, the background color
should be rendered inside the parent task. Also, there are some
CTS tests that verify that this surface is rendered within the
task when an activity-to-activity transition happens.

This CL limites the original logic only to the split case.
Having this logic for split is okay in freeform mode too.

Bug: 307705229
Test: atest ActivityTransitionTests
Change-Id: I94f0fb792998574fa9fbc36147d9b83af62a7f0a