Scan SQL comments when categorizing statements
This modifies the function that guesses at the SQL statement type by
having it scan past comments, if any are present.  Comments are
skipped using regular expressions, but only if the legacy statement
guess recognizes that a comment is present.  That is, there is no
performance penalty for statements that do not have leading comments.

A concurrency test has been added to verify that read-only statements
with comments are correctly assigned to non-primary database
connections and therefore can run in parallel with read/write
statements.

Test: atest
 * FrameworksCoreTests:android.database
 * CtsDatabaseTestCases

Bug: 270315305

Change-Id: Idf56f8ad279d54c701c0b53dc14a5879c8594f58