Ensure correct user id is attached to URIs.
When attempting to grant a URI permission, the user id for which the
grant was requested for was being omitted. Update the intent which is
passed to check the uri permission.

This was a regression caused by a no-op refactoring in
change Ied529156205903f9b02b4265963fdf59f7dd7f92.

Bug: 297302560
Test: atest ContextTest
Change-Id: Ibdd01c42a61ee53c15f322fea093a8d72ae3eedf