Tolerate "--dismiss-keyguard"
In UDC the flag was renamed to "--dismiss-keyguard" in a
change that has been reverted. Before and after, the
correct flag is "--dismiss-keyguard-if-insecure".

However, since UDC CTS passes in this flag, and must
pass against QPRs too, we need to accept this renamed
flag for at least the next year.

Change-Id: I5b4284a578ea917aef0304e05d27cbc512606d60
Test: Run UDC CTS KeyguardTests against 24Q1 Device
Fix: 310091307