Early return for task-start-position-only change
This CL makes containers in a task don't update if the task
is just moved but not resized. An example is drag-and drop
a freeform task. Containers don't need to update because
we use relative bounds in WCT operations. Relative bounds
won't be changed when a task is moved.
Note that we still need to check if the overlay container
should be dismissed in case the overlay container is the last
child window container of the parent task.

Test: atest OverlayPresentationTest
Test: atest WMJetpackUnitTests
Test: atest PinActivityStackTests
Bug: 243518738

Change-Id: Iaa6a48d8145caea42cf984b02cd4a7aafbfe96be