Excluding pip Task while trying to move Activity on Top
In ActivityEmbedding, touching on the embedded and non-current-focused
Activity will make the Activity focused if the parent Task is not moving
to top. However, it doesn't work when there's a pip Task. That's because
the pip Task is always-on-top and also focusable. The touched Task is
therefore not the top focusable Task and was requested to move to top
(though it is a no-op since it cannot be moved above pip).

Bug: 311984433
Test: ActivityRecordTests
Change-Id: Ib3f92d81b5a2daf9bb6dc4b7dbdfa53f9d45c7bd