Update mirroring dialog to show concurrent displays warning
This introduces config_concurrentDisplayDeviceStates array, that is device specific and is supposed to contain all device states that represent "concurrent displays".

This also creates a DeviceStateRepository that allows interactors to use the state provided by DeviceStateManager easily.
Several other places in sysui are doing something similar (e.g. DevicePostureController and DisplayStateRepository), but with a slighly different logic that doesn't suit this use case (DPC is using an androidx related res that doesn't contain the concurrent state and having some logic to use the base state in certain cases, and DSR is only listening for specific states). Eventually, those other classes should be refactored to use DeviceStateRepository under the wood.

This is only enabled for devices overriding the config_concurrentDisplayDeviceStates array in an overlay.

Flag: ACONFIG enable_dual_display_blocking DISABLED
Test: ConnectedDisplayInteractorTest, DeviceStateRepositoryTest, MirroringConfirmationDialogScerenshotTest
Bug: 296211844
Change-Id: I68c0b1489019471aec0a72fda70f57a7bc1ed29d