[flexiglass] Remove the redundant `isThrottled` state.
Currently, we have a state for `throttled` (which is meaningless when
throttling is inactive) and another one for `isThrottled`. This can lead
to inconsistent state and adds an unnecessary derived state to maintain.
This CL makes `throttled` null when throttling is inactive, removing the
needs for `isThrottled` to convey that.

Bonus: this converts the repository's `throttled` into a
`MutableStateFlow`, per the new repository conventions outlined in
go/repository-conventions.

Bug: 314757822
Test: Unit tests updated and still pass.
Flag: ACONFIG com.android.systemui.scene_container DEVELOPMENT
Change-Id: I4be0eddf83d2aad095ea0e291635ba1e459779fe