Stylus: Introduce Show Stylus Hover Pointer Setting (1/2)
Currently, the stylus hover icon can only be shown when
config_enableStylusPointerIcon is set to true. This means that the
users do not have a choice of turning the feature on or off.

This CL introduces the capability to do the same. This CL introduces
a new Secure Setting, STYLUS_POINTER_ICON_ENABLED which can be used to
turn on/off this feature.

Test: manual
Steps:
1. Change the value of config_enableStylusPointerIcon as true.
2. Build the images and flash it to a stylus supporting device.
3. After the device is booted, hover over the screen, the stylus
   pointer should appear.
4. Go to Settings > Connected Devices > Stylus and tap on
   "Show pointer while hovering" toggle.
5. Try hovering over the screen, check no pointer icon is shown
6. Repeat Step#4, check that stylus pointer appears.

Run
- make RunSettingsRoboTests \
      ROBOTEST_FILTER=StylusDevicesControllerTest
- atest SettingsProviderTest

Change-Id: Ie4a0e708ba2e5f9b95f9a4ebd96786e64d9c30cd