Fix SdkSandboxStorageHostTest
When the feature is enabled, the system doesn’t give app id if the
installed package is a sdk library. The Settings writePackageLPr()
will write the appIds to xml, it will read these appIds when booting.
The system will ignore appid -1 to create packageSetting that causes
the app dependent on this sdk-library be removed. The permission
service is also NPE because this app is removed.

Try to write a flag to help indentify if the system needs to create
PackageSettungs for sdk library.

Bug: 314772888
Test: atest SdkSandboxStorageHostTest
Change-Id: Ib53e775db123d0008948f1c2a1702bdaa0a47b76