Create cancellation signal closer to where face detection runs
With this change the detectCancellationSignal only gets modified from the main thread.
The nonNull check started appearing (rarely) out in the wild from 2023-12-05 a couple of days after we moved flows to the bg dispatcher, that meant this line was being executed in the background thread.
This error stopped in builds appearing created after 2023-12-13 after we reverted the change.

This change is a no-op right now

Bug: 316255379
Flag: NONE
Test: everything builds.
Change-Id: Ied8c792344a75329a799d53ebc586490f028d51c