Add timeout log if sync transaction is not applied
There are a few cases where the sync transaction is sent to WMS which
means it becomes outside the SSG control. In most cases, there's already
a timeout for BlastSyncEngine when sending the transaction back to
shell, but there are some exceptions.

This log will help identify if there are cases where something is
requesting to sync the buffer but isn't applying it.

Test: Builds, logs are not shown in most scenarios
Bug: 315285692
Change-Id: I2ec3fa5872b5c33277e4599aa1d3e355590ff70c