Offer `core-libart-for-host` under Ravenwood.
Ravenwood is a deviceless testing environment that runs in a host
JVM, similar to Robolectric.

As we're working on onboard existing SystemUI tests that expect
parity with Robolectric, we're needing to offer the same baseline
Robolectric functionality, so we expand the visibility of
`core-libart-for-host` to only the top-level of "//frameworks/base"
which is where Ravenwood.bp is defined.  (We're not exposing the
library to any subpackages.)

Bug: 319647875
Test: atest SystemUiRavenTests
Change-Id: I02185deb302aa764a6970de848d82788f7d07e19