Add navigation bar button logs
Add a historical log for when navigation bar buttons (recents, back, home, ime switch, accessibility) are pressed using log buffers to capture and dump this data. When phone starts and the buttons mentioned above are clicked, a historical log is maintained. To access the history of button clicks, run "adb shell dumpsys activity service SystemUIAuxiliaryDumpService" in the terminal.

Flag: NONE
Test: Manually tested in pixel 7 by pressing home, back, & recent
buttons in 3-button mode and checking for output.
Bug: 319682924
Change-Id: I423a8da101ebb7c28aff89e56f5050b3071f3972

Change-Id: I8f4c5f4aa3295900c88ce924fa7750974c7a7952