Add lift to type support
We already had existing life to type support, but only for the
numpadkeys. Using setTextEntryKey(true) in order to enable this feature.
There was note that if this is turned on, then the numpad key is clicked
twice. This is because, we already had a lift to type listener. I
removed this implementation in favor of the new one.

Test: Test pin screen with talkback on. Swipe finger to digits and
backspace to activate
Fixes: 54769853
Flag: NONE

Change-Id: Ie1630bf6fcea36e3cdcb8ee142f4339296c6843b