Migrate Bubbles to ProtoLog
Previously most of bubbles logging required you to rebuild and
flip flags in BubbleDebugConfig. Additionally, some of the logs were
a bit spammy and unhelpful.

This CL attempts to clean the logging up, add some additional logging,
and migrate most of the non-error related logs to ProtoLog.

I think there might still be some logs to prune / add but this should
be more useful than it was before.

Test: make && adb shell dumpsys activity service SystemUIService WMShell protolog enable-text WM_SHELL_BUBBLES
      and observe the logs
Bug: 323591549
Change-Id: If528dc0adcf5910fd62a4229777cfd391a2f7307