Vanquish KeyguardTransitionInteractorFactory.
This is no longer needed thanks to Kosmos.

This CL deletes the factory, introduces kosmos to the tests that used it, fixes tests where necessary, and introduces any needed kosmos fixtures.

Due to the size of this CL, I made as few changes as possible to each test to get them to pass. We'll make more passes through these tests to clean them up and also remove additional *Factory classes that have been introduced.

Bug: 278086361
Flag: NA
Test: atest SystemUITests
Change-Id: Ic8b03c5053436d6b05b805081b815fc00ff9b94c