Resolve API feedback TelecomManager#isInSelfManagedCall
As part of the mainline work, feedback has given to change the
hasCrossUserAccess param to better match the description. The
INTERACT_ACROSS_USERS permission has also been made an optional
parameter which is enforced when the user handle passed in doesn't match
the calling user or when the detectForAllUsers parameter is set to true.

Bug: 323958718
Bug: 311773409
Test: atest SelfManagedConnectionServiceTest
Change-Id: I96c04249e6320bde355a1efeddf0374048b84198