Move a11y window bounds computation into a11y package
This moves accessibility windows' touchable region computation and
exposed window selection from the wm package to the a11y package.
This shouldn't change any behavior. Our final goal is to eliminate the
dependency to WM package, but this just moves the code so that future
work will be smoother.

This change is guarded by a new flag.

Bug: 322444245
Test: CtsAccessibilityServiceTestCases CtsAccessibilityTestCases AccessibilityWindowManagerTest
Change-Id: Ib37cb934686d7677f2da74d6ed9ad6e7fcb5efd1