Fix IMMS critical dump winscope compatibility
Winscope expects the critical dump to be in the shape of a normal trace
with a single entry, however we currently only dump one nested field of
one entry message.

This fixes the compatibility by ensuring the required shape, enabling
better integration with winscope, and consistency with the critical
dumps from window manager and surface flinger.

Bug: 325016290
Test: n/a
Change-Id: I3c1f4f03cf043500c70fb7dddb80454979a07a80