Don't stop current user's profile even if excess
In Ie08342bb77544117cab4423b543af74cbeb109e2 we fixed some bugs relating
to stopping profiles' parents. However, this could cause the "force"
parameter to not be handled properly. We almost never actually want it,
but:

The case of stopping users when exceeding the maximum running users
is one of the few cases where we want "force==false" to truly not
stop a profile if its parent is the current user. As a result, we cannot
simply ignore the notion of "force".
We need to really refactor this notion of "force", but in the meantime,
here's a quick fix and test.

Test: atest UserControllerTest#testStoppingExcessRunningUsersAfterSwitch_currentProfileNotStopped
Bug: 324647580
Bug: 310249114
Change-Id: I4f9c5a11908fd9f7ad381d337d11af5c3058bc3d