TouchExplorer: Send hover events based on the event stream.
Previously there were a few areas where we sent hover events based on the internal state of TouchExplorer.
The issue is that when doing gesture detection in talkback, the state is no longer necessarily accurate in relation to the event stream.
The event stream is the source of truth for crashes.

Flag: send_hover_events_based_on_event_stream
Fix: 314251047
Test: atest FrameworksServicesTests:TouchExplorerTest CtsAccessibilityServiceTestCases:TouchExplorer
Test

Change-Id: Id34aaaea94b4ea81be9d4341761f9163a06f5865