Fix race condition in DREAM->LOCKSCREEN transition
When the dream ends, there is a race condition between the isDreaming
signal and the occluded signal. To ensure that we let occluded flip to
false first before dreaming does, in order to avoid erronously
transitioning to an OCCLUDED state, we add a small debounce to the
dreaming signal.

Flag: NA
Bug: 331240295
Test: verified DREAMING->LOCKSCREEN and DREAMING->OCCLUDED works
correctly by looking at KeyguardTransitionRepository logs

Change-Id: Ib5a1a61f89eed7f8541a1b2ce0d9e7a7344605e5