Fix wrongly applied material theme for StandardMenuPopup
The candidate windows has updated for Material theme but
it was done only for CascadingMenuPopup. The same things
need to be done for StandardMenuPopup which is mainly
used for the phone devices.

This CL also fixes the icon position wrongly aligned
to the left of the context menu for some case.

Bug: 332542108
Test: Manually tested on Pixel 8 and Pixel Tablet
Change-Id: I83918055f7ee1881061e6cf2744a578a838bebe5