Avoid nested tracing for showImePostLayout
When we have simultaenous show requests reaching
scheduleShowImePostLayout before either can continue, we create traces
for both using the same name and the same cookie, which results in
undefined behaviour for the tracing logic.

This moves the start of the WMS.showImePostLayout trace in
ImeInsetsSourceProvider, and makes sure we only start the trace once,
for the initial schedule call, measuring the end-to-end latency between
scheduling showing the IME post layout and succeeding / aborting.

Note this could slightly lower the values for this metric, as we start
measuring slightly later, skipping three calls in
WindowManagerService#showImePostLayout.

Bug: 332898042
Test: atest ImePerfTest
Change-Id: I4dae001a6030baf2a5339c0c1e8de33ccd89243a