Add audioserver to AppOps uid list
AppOps error logs "Operation not found" for finishOp for audioserver
(particularly WAKE_LOCK), which floods the logcat. This is because
appops doesn't track the UID state for audioserver, so any ops are
always MODE_ERRORED.

Add audioserver to the list of non-package UIDs to simulate uidstate
which allows for correct appops attribution.

Test: Manual, no errors in logcat
Bug: 333950879

Change-Id: I9ffea259c6b40a5c7246043bc1b0481429ce7948