Start using InputMethodSettingsRepository (2nd try)
This is the 2nd attempt of submitting my previous CL [1], which got
reverted due to a boottime regression (Bug 329414077). We have already
addressed the major contributor of the boot time regression [3], so
this CL should be ready to be submitted again.

With this CL we start taking advantage of

  InputMethodSettingsRepository,

which enables us to get a cached InputMethodSettings instead of
dynamically querying available IMEs.

To keep the CL as minimum as possible,

  InputMethodManagerService#mSettings

remains unchanged. It will be removed in a subsequent bug 329163064.

There must be no user observable behavior change.

 [1]: Ie2d3b7e9b75790f64789d0c2e96ae46b1dee3a1e
      ba0d697ba66efe7465eec75ee055c2a375fd19a2
 [2]: I20317bdd012c51a5ba57cd6e4d2e1a69df66d6a2
      bb339d01d54deba5fff1b47805877eec6a56c56f
 [3]: I8e51b1ced4dc16cdca7e898885c64793665fafef
      10ae88b322245aacbbfeed32097206f682f75a35

Bug: 329163064
Bug: 333591545
Fix: 309837937
Test: presubmit
Change-Id: Ic0dd655fbd86b8ccce2b3298b4c70359a468f9ec