Fix test-suite for test_module_configs
With upcoming changes, test_module_config modules need to live in
the same test-suite as their base, in this case FrameworksServicesTests.
It lives in device-tests, so the derived tests should as well.

In the future we can migrate to general-tests if that makes more sense.
FYI: b/149420197 has some context on the difference.

Test: mm
Test: atest FrameworksServicesTests_contentprotection
Change-Id: I8358e2b026eea4b8fe449e9c51d13ea5623389a1