When resizing split and both sides are veiled, make sure both sides are veiled with the same color
This is a temporary quick fix for b/339274381. When a split resize is initiated and both sides are being veiled (like when Chrome is opening a new instance of itself), there is a bug where the two sides return different colors, which looks bad. This patch ensures that both sides are veiled with the same color.

When b/339274381 is fixed, this CL can be reverted entirely.

Bug: 307490004
Test: Launch an adjacent task from Chrome using the repro steps in b/339274381, see that both sides are the same color.
Flag: None
Change-Id: I3158a1da29ec58c5412401b67a6905c31bb120b1
Merged-In: I3158a1da29ec58c5412401b67a6905c31bb120b1