Use a dedicated thread for IMMS.MyPackageMonitor
This is a follow up CL to my previous CL [1], which moved resource I/O
out of ImfLock to reduce the risk of ANRs.

Once thing that is still concerning is that

  MyPackageMonitor#onFinishPackageChangesInternal()

still gets called on a worker thread that is shared across multiple
operations in InputMethodManagerService, e.g.

 * HandwritingModeController as a UI event handler
 * event callbacks for ContentObserver
 * event forwarding used by ZeroJankProxy

Blocking above operations with resource I/O is also something we would
like to avoid. Thus this CL introduces a new worker thread dedicated
for MyPackageMonitor.

There should be no observable semantic change.

 [1]: I263cd49dd4d64b64136acc3dad469f83a862ce97
      f70ec6b1d7a2bdd69fe6a721fa65731fab3b4881

Bug: 340221861
Test: presubmit
Change-Id: Icc84fe9d0d3b5842b4d5ee51b12b0d3cd10a1d24